@@ -25,7 +25,7 @@ const (
 // This struct should be used by generated code only.
 type Message struct {
 	Metadata *ordering.Metadata
-	Message  protoreflect.ProtoMessage
+	Message  proto.Message
 	msgType  gorumsMsgType
 }
 
@@ -50,20 +50,20 @@ func NewCodec() *Codec {
 }
 
 // Name returns the name of the Codec.
-func (c Codec) Name() string {
+func (Codec) Name() string {
 	return ContentSubtype
 }
 
-func (c Codec) String() string {
+func (Codec) String() string {
 	return ContentSubtype
 }
 
 // Marshal marshals the message m into a byte slice.
-func (c Codec) Marshal(m interface{}) (b []byte, err error) {
+func (c Codec) Marshal(m any) (b []byte, err error) {
 	switch msg := m.(type) {
 	case *Message:
 		return c.gorumsMarshal(msg)
-	case protoreflect.ProtoMessage:
+	case proto.Message:
 		return c.marshaler.Marshal(msg)
 	default:
 		return nil, fmt.Errorf("gorums: cannot marshal message of type '%T'", m)
@@ -89,11 +89,11 @@ func (c Codec) gorumsMarshal(msg *Message) (b []byte, err error) {
 }
 
 // Unmarshal unmarshals a byte slice into m.
-func (c Codec) Unmarshal(b []byte, m interface{}) (err error) {
+func (c Codec) Unmarshal(b []byte, m any) (err error) {
 	switch msg := m.(type) {
 	case *Message:
 		return c.gorumsUnmarshal(b, msg)
-	case protoreflect.ProtoMessage:
+	case proto.Message:
 		return c.unmarshaler.Unmarshal(b, msg)
 	default:
 		return fmt.Errorf("gorums: cannot unmarshal message of type '%T'", m)
@@ -106,13 +106,14 @@ func (c Codec) gorumsUnmarshal(b []byte, msg *Message) (err error) {
 	mdBuf, mdLen := protowire.ConsumeBytes(b)
 	err = c.unmarshaler.Unmarshal(mdBuf, msg.Metadata)
 	if err != nil {
-		return err
+		return fmt.Errorf("gorums: could not unmarshal metadata: %w", err)
 	}
 
 	// get method descriptor from registry
 	desc, err := protoregistry.GlobalFiles.FindDescriptorByName(protoreflect.FullName(msg.Metadata.GetMethod()))
 	if err != nil {
-		return err
+		// err is a NotFound error with no method name information; return a more informative error
+		return fmt.Errorf("gorums: could not find method descriptor for %s", msg.Metadata.GetMethod())
 	}
 	methodDesc := desc.(protoreflect.MethodDescriptor)
 
@@ -130,7 +131,8 @@ func (c Codec) gorumsUnmarshal(b []byte, msg *Message) (err error) {
 	// now get the message type from the types registry
 	msgType, err := protoregistry.GlobalTypes.FindMessageByName(messageName)
 	if err != nil {
-		return err
+		// err is a NotFound error with no message name information; return a more informative error
+		return fmt.Errorf("gorums: could not find message type %s", messageName)
 	}
 	msg.Message = msgType.New().Interface()
 
