@@ -14,8 +14,8 @@ import (
 // n is the number of replicas, and div is a divider.
 // the target level is n, and the level is calculated by the quorum function
 // by dividing the sum of levels from the servers with the divider.
-func run(t *testing.T, n int, div int, corr func(context.Context, *Configuration) *gorums.Correctable) {
-	addrs, teardown := gorums.TestSetup(t, n, func(i int) gorums.ServerIface {
+func run(t *testing.T, n, div int, corr func(context.Context, *Configuration, int, int) *gorums.Correctable[int]) {
+	addrs, teardown := gorums.TestSetup(t, n, func(_ int) gorums.ServerIface {
 		gorumsSrv := gorums.NewServer()
 		RegisterCorrectableTestServer(gorumsSrv, &testSrv{n})
 		return gorumsSrv
@@ -23,7 +23,6 @@ func run(t *testing.T, n int, div int, corr func(context.Context, *Configuration
 	defer teardown()
 
 	cfg, err := NewConfiguration(
-		qspec{div, n},
 		gorums.WithNodeList(addrs),
 		gorums.WithGrpcDialOptions(
 			grpc.WithTransportCredentials(insecure.NewCredentials()),
@@ -35,7 +34,7 @@ func run(t *testing.T, n int, div int, corr func(context.Context, *Configuration
 
 	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
 	defer cancel()
-	res := corr(ctx, cfg)
+	res := corr(ctx, cfg, n, div)
 
 	donech := res.Done()
 
@@ -54,59 +53,51 @@ func run(t *testing.T, n int, div int, corr func(context.Context, *Configuration
 	}
 }
 
-func TestCorrectable(t *testing.T) {
-	run(t, 4, 1, func(ctx context.Context, c *Configuration) *gorums.Correctable {
-		corr := c.Correctable(ctx, &CorrectableRequest{})
-		return corr.Correctable
-	})
-}
-
-func TestCorrectableStream(t *testing.T) {
-	run(t, 4, 4, func(ctx context.Context, c *Configuration) *gorums.Correctable {
-		corr := c.CorrectableStream(ctx, &CorrectableRequest{})
-		return corr.Correctable
-	})
-}
-
 type qspec struct {
 	div       int
 	doneLevel int
 }
 
-func (q qspec) q(replies map[uint32]*CorrectableResponse) (*CorrectableResponse, int, bool) {
+func (q qspec) corrQF(responses gorums.Responses[*CorrectableResponse], levelSet func(int, int, error)) {
 	sum := 0
-	for _, reply := range replies {
-		sum += int(reply.Level)
+	for response := range responses {
+		msg, err, _ := response.Unpack()
+		if err != nil {
+			levelSet(0, 0, err)
+			continue
+		}
+		sum += int(msg.GetLevel())
+		level := sum / q.div
+		levelSet(level, level, nil)
+		if level >= q.doneLevel {
+			return
+		}
 	}
-	level := sum / q.div
-	return &CorrectableResponse{Level: int32(level)}, level, level >= q.doneLevel
 }
 
-// CorrectableStreamQF is the quorum function for the CorrectableStream
-// correctable stream quorum call method. The in parameter is the request object
-// supplied to the CorrectableStream method at call time, and may or may not
-// be used by the quorum function. If the in parameter is not needed
-// you should implement your quorum function with '_ *CorrectableRequest'.
-func (q qspec) CorrectableStreamQF(_ *CorrectableRequest, replies map[uint32]*CorrectableResponse) (*CorrectableResponse, int, bool) {
-	return q.q(replies)
+func TestCorrectable(t *testing.T) {
+	run(t, 4, 1, func(ctx context.Context, c *Configuration, n, div int) *gorums.Correctable[int] {
+		corr := c.Correctable(ctx, &CorrectableRequest{})
+		qspec := qspec{div, n}
+		return gorums.NewCorrectable(corr, qspec.corrQF)
+	})
 }
 
-// CorrectableQF is the quorum function for the Correctable
-// correctable quorum call method. The in parameter is the request object
-// supplied to the Correctable method at call time, and may or may not
-// be used by the quorum function. If the in parameter is not needed
-// you should implement your quorum function with '_ *CorrectableRequest'.
-func (q qspec) CorrectableQF(_ *CorrectableRequest, replies map[uint32]*CorrectableResponse) (*CorrectableResponse, int, bool) {
-	return q.q(replies)
+func TestCorrectableStream(t *testing.T) {
+	run(t, 4, 4, func(ctx context.Context, c *Configuration, n, div int) *gorums.Correctable[int] {
+		corr := c.CorrectableStream(ctx, &CorrectableRequest{})
+		qspec := qspec{div, n}
+		return gorums.NewCorrectable(corr, qspec.corrQF)
+	})
 }
 
 type testSrv struct {
 	n int
 }
 
-func (srv testSrv) CorrectableStream(_ gorums.ServerCtx, request *CorrectableRequest, send func(response *CorrectableResponse) error) error {
-	for i := 0; i < srv.n; i++ {
-		err := send(&CorrectableResponse{Level: int32(i + 1)})
+func (srv testSrv) CorrectableStream(_ gorums.ServerCtx, _ *CorrectableRequest, send func(response *CorrectableResponse) error) error {
+	for i := range srv.n {
+		err := send(CorrectableResponse_builder{Level: int32(i + 1)}.Build())
 		if err != nil {
 			return err
 		}
@@ -114,6 +105,6 @@ func (srv testSrv) CorrectableStream(_ gorums.ServerCtx, request *CorrectableReq
 	return nil
 }
 
-func (srv testSrv) Correctable(_ gorums.ServerCtx, request *CorrectableRequest) (response *CorrectableResponse, err error) {
-	return &CorrectableResponse{Level: 1}, nil
+func (testSrv) Correctable(_ gorums.ServerCtx, _ *CorrectableRequest) (response *CorrectableResponse, err error) {
+	return CorrectableResponse_builder{Level: 1}.Build(), nil
 }
