@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.5
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: benchmark/benchmark.proto
 
 package benchmark
@@ -527,187 +527,85 @@ func (b0 MemoryStat_builder) Build() *MemoryStat {
 	return m0
 }
 
-type MemoryStatList struct {
-	state                  protoimpl.MessageState `protogen:"opaque.v1"`
-	xxx_hidden_MemoryStats *[]*MemoryStat         `protobuf:"bytes,1,rep,name=MemoryStats"`
-	unknownFields          protoimpl.UnknownFields
-	sizeCache              protoimpl.SizeCache
-}
-
-func (x *MemoryStatList) Reset() {
-	*x = MemoryStatList{}
-	mi := &file_benchmark_benchmark_proto_msgTypes[7]
-	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-	ms.StoreMessageInfo(mi)
-}
-
-func (x *MemoryStatList) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*MemoryStatList) ProtoMessage() {}
-
-func (x *MemoryStatList) ProtoReflect() protoreflect.Message {
-	mi := &file_benchmark_benchmark_proto_msgTypes[7]
-	if x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-func (x *MemoryStatList) GetMemoryStats() []*MemoryStat {
-	if x != nil {
-		if x.xxx_hidden_MemoryStats != nil {
-			return *x.xxx_hidden_MemoryStats
-		}
-	}
-	return nil
-}
-
-func (x *MemoryStatList) SetMemoryStats(v []*MemoryStat) {
-	x.xxx_hidden_MemoryStats = &v
-}
-
-type MemoryStatList_builder struct {
-	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
-
-	MemoryStats []*MemoryStat
-}
-
-func (b0 MemoryStatList_builder) Build() *MemoryStatList {
-	m0 := &MemoryStatList{}
-	b, x := &b0, m0
-	_, _ = b, x
-	x.xxx_hidden_MemoryStats = &b.MemoryStats
-	return m0
-}
-
 var File_benchmark_benchmark_proto protoreflect.FileDescriptor
 
-var file_benchmark_benchmark_proto_rawDesc = string([]byte{
-	0x0a, 0x19, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2f, 0x62, 0x65, 0x6e, 0x63,
-	0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x09, 0x62, 0x65, 0x6e,
-	0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x1a, 0x0c, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x22, 0x20, 0x0a, 0x04, 0x45, 0x63, 0x68, 0x6f, 0x12, 0x18, 0x0a, 0x07, 0x70, 0x61, 0x79,
-	0x6c, 0x6f, 0x61, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x07, 0x70, 0x61, 0x79, 0x6c,
-	0x6f, 0x61, 0x64, 0x22, 0x40, 0x0a, 0x08, 0x54, 0x69, 0x6d, 0x65, 0x64, 0x4d, 0x73, 0x67, 0x12,
-	0x1a, 0x0a, 0x08, 0x53, 0x65, 0x6e, 0x64, 0x54, 0x69, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x03, 0x52, 0x08, 0x53, 0x65, 0x6e, 0x64, 0x54, 0x69, 0x6d, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x70,
-	0x61, 0x79, 0x6c, 0x6f, 0x61, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x07, 0x70, 0x61,
-	0x79, 0x6c, 0x6f, 0x61, 0x64, 0x22, 0x0e, 0x0a, 0x0c, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x0f, 0x0a, 0x0d, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x0d, 0x0a, 0x0b, 0x53, 0x74, 0x6f, 0x70, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0xad, 0x02, 0x0a, 0x06, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74,
-	0x12, 0x12, 0x0a, 0x04, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04,
-	0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x4f, 0x70, 0x73,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x08, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x4f, 0x70, 0x73,
-	0x12, 0x1c, 0x0a, 0x09, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x54, 0x69, 0x6d, 0x65, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x03, 0x52, 0x09, 0x54, 0x6f, 0x74, 0x61, 0x6c, 0x54, 0x69, 0x6d, 0x65, 0x12, 0x1e,
-	0x0a, 0x0a, 0x54, 0x68, 0x72, 0x6f, 0x75, 0x67, 0x68, 0x70, 0x75, 0x74, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x01, 0x52, 0x0a, 0x54, 0x68, 0x72, 0x6f, 0x75, 0x67, 0x68, 0x70, 0x75, 0x74, 0x12, 0x1e,
-	0x0a, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x41, 0x76, 0x67, 0x18, 0x05, 0x20, 0x01,
-	0x28, 0x01, 0x52, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x41, 0x76, 0x67, 0x12, 0x1e,
-	0x0a, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x56, 0x61, 0x72, 0x18, 0x06, 0x20, 0x01,
-	0x28, 0x01, 0x52, 0x0a, 0x4c, 0x61, 0x74, 0x65, 0x6e, 0x63, 0x79, 0x56, 0x61, 0x72, 0x12, 0x20,
-	0x0a, 0x0b, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x50, 0x65, 0x72, 0x4f, 0x70, 0x18, 0x07, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x0b, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x50, 0x65, 0x72, 0x4f, 0x70,
-	0x12, 0x1a, 0x0a, 0x08, 0x4d, 0x65, 0x6d, 0x50, 0x65, 0x72, 0x4f, 0x70, 0x18, 0x08, 0x20, 0x01,
-	0x28, 0x04, 0x52, 0x08, 0x4d, 0x65, 0x6d, 0x50, 0x65, 0x72, 0x4f, 0x70, 0x12, 0x37, 0x0a, 0x0b,
-	0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x53, 0x74, 0x61, 0x74, 0x73, 0x18, 0x09, 0x20, 0x03, 0x28,
-	0x0b, 0x32, 0x15, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x4d, 0x65,
-	0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61, 0x74, 0x52, 0x0b, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72,
-	0x53, 0x74, 0x61, 0x74, 0x73, 0x22, 0x3c, 0x0a, 0x0a, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53,
-	0x74, 0x61, 0x74, 0x12, 0x16, 0x0a, 0x06, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x06, 0x41, 0x6c, 0x6c, 0x6f, 0x63, 0x73, 0x12, 0x16, 0x0a, 0x06, 0x4d,
-	0x65, 0x6d, 0x6f, 0x72, 0x79, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x06, 0x4d, 0x65, 0x6d,
-	0x6f, 0x72, 0x79, 0x22, 0x49, 0x0a, 0x0e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61,
-	0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x37, 0x0a, 0x0b, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53,
-	0x74, 0x61, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x62, 0x65, 0x6e,
-	0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61,
-	0x74, 0x52, 0x0b, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61, 0x74, 0x73, 0x32, 0xb2,
-	0x04, 0x0a, 0x09, 0x42, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x12, 0x4f, 0x0a, 0x14,
-	0x53, 0x74, 0x61, 0x72, 0x74, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x42, 0x65, 0x6e, 0x63, 0x68,
-	0x6d, 0x61, 0x72, 0x6b, 0x12, 0x17, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b,
-	0x2e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x18, 0x2e,
-	0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04, 0xa0, 0xb5, 0x18, 0x01, 0x12, 0x46, 0x0a,
-	0x13, 0x53, 0x74, 0x6f, 0x70, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x42, 0x65, 0x6e, 0x63, 0x68,
-	0x6d, 0x61, 0x72, 0x6b, 0x12, 0x16, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b,
-	0x2e, 0x53, 0x74, 0x6f, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x11, 0x2e, 0x62,
-	0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x22,
-	0x04, 0xa0, 0xb5, 0x18, 0x01, 0x12, 0x49, 0x0a, 0x0e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x42, 0x65,
-	0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x12, 0x17, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d,
-	0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x1a, 0x18, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74, 0x61,
-	0x72, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04, 0xa0, 0xb5, 0x18, 0x01,
-	0x12, 0x56, 0x0a, 0x0d, 0x53, 0x74, 0x6f, 0x70, 0x42, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72,
-	0x6b, 0x12, 0x16, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x53, 0x74,
-	0x6f, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x15, 0x2e, 0x62, 0x65, 0x6e, 0x63,
-	0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x53, 0x74, 0x61, 0x74,
-	0x22, 0x16, 0xa0, 0xb5, 0x18, 0x01, 0xf2, 0xb6, 0x18, 0x0e, 0x4d, 0x65, 0x6d, 0x6f, 0x72, 0x79,
-	0x53, 0x74, 0x61, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x34, 0x0a, 0x0a, 0x51, 0x75, 0x6f, 0x72,
-	0x75, 0x6d, 0x43, 0x61, 0x6c, 0x6c, 0x12, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61,
-	0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x1a, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d,
-	0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x22, 0x04, 0xa0, 0xb5, 0x18, 0x01, 0x12, 0x3d,
-	0x0a, 0x0f, 0x41, 0x73, 0x79, 0x6e, 0x63, 0x51, 0x75, 0x6f, 0x72, 0x75, 0x6d, 0x43, 0x61, 0x6c,
-	0x6c, 0x12, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63,
-	0x68, 0x6f, 0x1a, 0x0f, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45,
-	0x63, 0x68, 0x6f, 0x22, 0x08, 0xa0, 0xb5, 0x18, 0x01, 0xd0, 0xb5, 0x18, 0x01, 0x12, 0x34, 0x0a,
-	0x0a, 0x53, 0x6c, 0x6f, 0x77, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x12, 0x0f, 0x2e, 0x62, 0x65,
-	0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x1a, 0x0f, 0x2e, 0x62,
-	0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x45, 0x63, 0x68, 0x6f, 0x22, 0x04, 0xa0,
-	0xb5, 0x18, 0x01, 0x12, 0x3e, 0x0a, 0x09, 0x4d, 0x75, 0x6c, 0x74, 0x69, 0x63, 0x61, 0x73, 0x74,
-	0x12, 0x13, 0x2e, 0x62, 0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x2e, 0x54, 0x69, 0x6d,
-	0x65, 0x64, 0x4d, 0x73, 0x67, 0x1a, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x04, 0x98,
-	0xb5, 0x18, 0x01, 0x42, 0x28, 0x5a, 0x21, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f,
-	0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x62,
-	0x65, 0x6e, 0x63, 0x68, 0x6d, 0x61, 0x72, 0x6b, 0x92, 0x03, 0x02, 0x08, 0x02, 0x62, 0x08, 0x65,
-	0x64, 0x69, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x70, 0xe8, 0x07,
-})
+const file_benchmark_benchmark_proto_rawDesc = "" +
+	"\n" +
+	"\x19benchmark/benchmark.proto\x12\tbenchmark\x1a\x1bgoogle/protobuf/empty.proto\x1a\fgorums.proto\" \n" +
+	"\x04Echo\x12\x18\n" +
+	"\apayload\x18\x01 \x01(\fR\apayload\"@\n" +
+	"\bTimedMsg\x12\x1a\n" +
+	"\bSendTime\x18\x01 \x01(\x03R\bSendTime\x12\x18\n" +
+	"\apayload\x18\x02 \x01(\fR\apayload\"\x0e\n" +
+	"\fStartRequest\"\x0f\n" +
+	"\rStartResponse\"\r\n" +
+	"\vStopRequest\"\xad\x02\n" +
+	"\x06Result\x12\x12\n" +
+	"\x04Name\x18\x01 \x01(\tR\x04Name\x12\x1a\n" +
+	"\bTotalOps\x18\x02 \x01(\x04R\bTotalOps\x12\x1c\n" +
+	"\tTotalTime\x18\x03 \x01(\x03R\tTotalTime\x12\x1e\n" +
+	"\n" +
+	"Throughput\x18\x04 \x01(\x01R\n" +
+	"Throughput\x12\x1e\n" +
+	"\n" +
+	"LatencyAvg\x18\x05 \x01(\x01R\n" +
+	"LatencyAvg\x12\x1e\n" +
+	"\n" +
+	"LatencyVar\x18\x06 \x01(\x01R\n" +
+	"LatencyVar\x12 \n" +
+	"\vAllocsPerOp\x18\a \x01(\x04R\vAllocsPerOp\x12\x1a\n" +
+	"\bMemPerOp\x18\b \x01(\x04R\bMemPerOp\x127\n" +
+	"\vServerStats\x18\t \x03(\v2\x15.benchmark.MemoryStatR\vServerStats\"<\n" +
+	"\n" +
+	"MemoryStat\x12\x16\n" +
+	"\x06Allocs\x18\x01 \x01(\x04R\x06Allocs\x12\x16\n" +
+	"\x06Memory\x18\x02 \x01(\x04R\x06Memory2\x9c\x04\n" +
+	"\tBenchmark\x12O\n" +
+	"\x14StartServerBenchmark\x12\x17.benchmark.StartRequest\x1a\x18.benchmark.StartResponse\"\x04\xa0\xb5\x18\x01\x12F\n" +
+	"\x13StopServerBenchmark\x12\x16.benchmark.StopRequest\x1a\x11.benchmark.Result\"\x04\xa0\xb5\x18\x01\x12I\n" +
+	"\x0eStartBenchmark\x12\x17.benchmark.StartRequest\x1a\x18.benchmark.StartResponse\"\x04\xa0\xb5\x18\x01\x12D\n" +
+	"\rStopBenchmark\x12\x16.benchmark.StopRequest\x1a\x15.benchmark.MemoryStat\"\x04\xa0\xb5\x18\x01\x124\n" +
+	"\n" +
+	"QuorumCall\x12\x0f.benchmark.Echo\x1a\x0f.benchmark.Echo\"\x04\xa0\xb5\x18\x01\x129\n" +
+	"\x0fAsyncQuorumCall\x12\x0f.benchmark.Echo\x1a\x0f.benchmark.Echo\"\x04\xa0\xb5\x18\x01\x124\n" +
+	"\n" +
+	"SlowServer\x12\x0f.benchmark.Echo\x1a\x0f.benchmark.Echo\"\x04\xa0\xb5\x18\x01\x12>\n" +
+	"\tMulticast\x12\x13.benchmark.TimedMsg\x1a\x16.google.protobuf.Empty\"\x04\x98\xb5\x18\x01B(Z!github.com/relab/gorums/benchmark\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
-var file_benchmark_benchmark_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
+var file_benchmark_benchmark_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
 var file_benchmark_benchmark_proto_goTypes = []any{
-	(*Echo)(nil),           // 0: benchmark.Echo
-	(*TimedMsg)(nil),       // 1: benchmark.TimedMsg
-	(*StartRequest)(nil),   // 2: benchmark.StartRequest
-	(*StartResponse)(nil),  // 3: benchmark.StartResponse
-	(*StopRequest)(nil),    // 4: benchmark.StopRequest
-	(*Result)(nil),         // 5: benchmark.Result
-	(*MemoryStat)(nil),     // 6: benchmark.MemoryStat
-	(*MemoryStatList)(nil), // 7: benchmark.MemoryStatList
-	(*emptypb.Empty)(nil),  // 8: google.protobuf.Empty
+	(*Echo)(nil),          // 0: benchmark.Echo
+	(*TimedMsg)(nil),      // 1: benchmark.TimedMsg
+	(*StartRequest)(nil),  // 2: benchmark.StartRequest
+	(*StartResponse)(nil), // 3: benchmark.StartResponse
+	(*StopRequest)(nil),   // 4: benchmark.StopRequest
+	(*Result)(nil),        // 5: benchmark.Result
+	(*MemoryStat)(nil),    // 6: benchmark.MemoryStat
+	(*emptypb.Empty)(nil), // 7: google.protobuf.Empty
 }
 var file_benchmark_benchmark_proto_depIdxs = []int32{
-	6,  // 0: benchmark.Result.ServerStats:type_name -> benchmark.MemoryStat
-	6,  // 1: benchmark.MemoryStatList.MemoryStats:type_name -> benchmark.MemoryStat
-	2,  // 2: benchmark.Benchmark.StartServerBenchmark:input_type -> benchmark.StartRequest
-	4,  // 3: benchmark.Benchmark.StopServerBenchmark:input_type -> benchmark.StopRequest
-	2,  // 4: benchmark.Benchmark.StartBenchmark:input_type -> benchmark.StartRequest
-	4,  // 5: benchmark.Benchmark.StopBenchmark:input_type -> benchmark.StopRequest
-	0,  // 6: benchmark.Benchmark.QuorumCall:input_type -> benchmark.Echo
-	0,  // 7: benchmark.Benchmark.AsyncQuorumCall:input_type -> benchmark.Echo
-	0,  // 8: benchmark.Benchmark.SlowServer:input_type -> benchmark.Echo
-	1,  // 9: benchmark.Benchmark.Multicast:input_type -> benchmark.TimedMsg
-	3,  // 10: benchmark.Benchmark.StartServerBenchmark:output_type -> benchmark.StartResponse
-	5,  // 11: benchmark.Benchmark.StopServerBenchmark:output_type -> benchmark.Result
-	3,  // 12: benchmark.Benchmark.StartBenchmark:output_type -> benchmark.StartResponse
-	6,  // 13: benchmark.Benchmark.StopBenchmark:output_type -> benchmark.MemoryStat
-	0,  // 14: benchmark.Benchmark.QuorumCall:output_type -> benchmark.Echo
-	0,  // 15: benchmark.Benchmark.AsyncQuorumCall:output_type -> benchmark.Echo
-	0,  // 16: benchmark.Benchmark.SlowServer:output_type -> benchmark.Echo
-	8,  // 17: benchmark.Benchmark.Multicast:output_type -> google.protobuf.Empty
-	10, // [10:18] is the sub-list for method output_type
-	2,  // [2:10] is the sub-list for method input_type
-	2,  // [2:2] is the sub-list for extension type_name
-	2,  // [2:2] is the sub-list for extension extendee
-	0,  // [0:2] is the sub-list for field type_name
+	6, // 0: benchmark.Result.ServerStats:type_name -> benchmark.MemoryStat
+	2, // 1: benchmark.Benchmark.StartServerBenchmark:input_type -> benchmark.StartRequest
+	4, // 2: benchmark.Benchmark.StopServerBenchmark:input_type -> benchmark.StopRequest
+	2, // 3: benchmark.Benchmark.StartBenchmark:input_type -> benchmark.StartRequest
+	4, // 4: benchmark.Benchmark.StopBenchmark:input_type -> benchmark.StopRequest
+	0, // 5: benchmark.Benchmark.QuorumCall:input_type -> benchmark.Echo
+	0, // 6: benchmark.Benchmark.AsyncQuorumCall:input_type -> benchmark.Echo
+	0, // 7: benchmark.Benchmark.SlowServer:input_type -> benchmark.Echo
+	1, // 8: benchmark.Benchmark.Multicast:input_type -> benchmark.TimedMsg
+	3, // 9: benchmark.Benchmark.StartServerBenchmark:output_type -> benchmark.StartResponse
+	5, // 10: benchmark.Benchmark.StopServerBenchmark:output_type -> benchmark.Result
+	3, // 11: benchmark.Benchmark.StartBenchmark:output_type -> benchmark.StartResponse
+	6, // 12: benchmark.Benchmark.StopBenchmark:output_type -> benchmark.MemoryStat
+	0, // 13: benchmark.Benchmark.QuorumCall:output_type -> benchmark.Echo
+	0, // 14: benchmark.Benchmark.AsyncQuorumCall:output_type -> benchmark.Echo
+	0, // 15: benchmark.Benchmark.SlowServer:output_type -> benchmark.Echo
+	7, // 16: benchmark.Benchmark.Multicast:output_type -> google.protobuf.Empty
+	9, // [9:17] is the sub-list for method output_type
+	1, // [1:9] is the sub-list for method input_type
+	1, // [1:1] is the sub-list for extension type_name
+	1, // [1:1] is the sub-list for extension extendee
+	0, // [0:1] is the sub-list for field type_name
 }
 
 func init() { file_benchmark_benchmark_proto_init() }
@@ -721,7 +619,7 @@ func file_benchmark_benchmark_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: unsafe.Slice(unsafe.StringData(file_benchmark_benchmark_proto_rawDesc), len(file_benchmark_benchmark_proto_rawDesc)),
 			NumEnums:      0,
-			NumMessages:   8,
+			NumMessages:   7,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
