@@ -1,92 +1,76 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+125b0cfb
+// 	protoc            v6.30.2
 // source: benchmark/benchmark.proto
 
 package benchmark
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
 // procedure calls may be invoked.
 type Configuration struct {
 	gorums.RawConfiguration
-	qspec QuorumSpec
-	nodes []*Node
 }
 
-func NewConfiguration(qspec QuorumSpec, cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *Configuration, err error) {
-	c = &Configuration{
-		qspec: qspec,
-	}
+// NewConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *Configuration, err error) {
+	c = &Configuration{}
 	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
 	if err != nil {
 		return nil, err
 	}
-	c.nodes = make([]*Node, c.Size())
-	for i, n := range c.RawConfiguration.RawNodes {
-		c.nodes[i] = &Node{n}
-	}
 	return c, nil
 }
 
-// ConfigurationFromRaw returns a new Configuration from the given raw configuration and QuorumSpec.
+// ConfigurationFromRaw returns a new configuration from the given raw configuration.
 //
-// This function may for example be used to "clone" a configuration but install a different QuorumSpec:
-//
-//	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
-//	cfg2 := ConfigurationFromRaw(cfg1.RawConfiguration, qspec2)
-func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
-	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
-	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+//	cfg1, err := pb.NewConfiguration(nodeList, opts...)
+//	cfg2 := ConfigurationFromRaw(cfg1.RawConfig)
+func ConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*Configuration, error) {
 	newCfg := &Configuration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*Node, newCfg.Size())
-	for i, n := range rawCfg.Nodes() {
-		newCfg.nodes[i] = &Node{n}
 	}
 	return newCfg, nil
 }
 
-func (c *Configuration) SubConfiguration(qspec QuorumSpec, cfg gorums.NodeListOption) (subCfg *Configuration, err error) {
-	subCfg = &Configuration{
-		qspec: qspec,
-	}
+// SubConfiguration allows for making a new Configuration from the ManagerOption list and
+// node list of another configuration,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *Configuration) SubConfiguration(cfg gorums.NodeListOption) (subCfg *Configuration, err error) {
+	subCfg = &Configuration{}
 	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
 	if err != nil {
 		return nil, err
-	}
-	subCfg.nodes = make([]*Node, subCfg.Size())
-	for i, n := range subCfg.RawConfiguration.Nodes() {
-		subCfg.nodes[i] = &Node{n}
 	}
 	return subCfg, nil
 }
 
 // Close closes a configuration created from the NewConfiguration method
 //
-// NOTE: A configuration created with ConfigurationFromRaw is closed when the original configuration is closed
+// NOTE: A configuration created with ConfigurationFromRaw or SubConfiguration closes and
+// is closed when the original configuration or any of the subconfigurations are closed.
 // If you want the configurations to be independent you need to use NewConfiguration
 func (c *Configuration) Close() error {
 	return c.RawConfiguration.Close()
@@ -97,17 +81,21 @@ func (c *Configuration) Close() error {
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *Configuration) Nodes() []*Node {
-	return c.nodes
+	nodes := make([]*Node, c.Size())
+	for i, n := range c.RawConfiguration.RawNodes {
+		nodes[i] = &Node{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
-func (c Configuration) And(d *Configuration) gorums.NodeListOption {
+func (c *Configuration) And(d *Configuration) gorums.NodeListOption {
 	return c.RawConfiguration.And(d.RawConfiguration)
 }
 
 // Except returns a NodeListOption that can be used to create a new configuration
 // from c without the nodes in rm.
-func (c Configuration) Except(rm *Configuration) gorums.NodeListOption {
+func (c *Configuration) Except(rm *Configuration) gorums.NodeListOption {
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
@@ -123,35 +111,15 @@ type Node struct {
 	*gorums.RawNode
 }
 
-// AsyncQuorumCall asynchronously invokes a quorum call on configuration c
-// and returns a AsyncEcho, which can be used to inspect the quorum call
-// reply and error when available.
-func (c *Configuration) AsyncQuorumCall(ctx context.Context, in *Echo) *AsyncEcho {
-	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.AsyncQuorumCall",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Echo, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Echo)
-		}
-		return c.qspec.AsyncQuorumCallQF(req.(*Echo), r)
-	}
-
-	fut := c.RawConfiguration.AsyncCall(ctx, cd)
-	return &AsyncEcho{fut}
-}
-
 // BenchmarkClient is the client interface for the Benchmark service.
 type BenchmarkClient interface {
-	StartServerBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error)
-	StopServerBenchmark(ctx context.Context, in *StopRequest) (resp *Result, err error)
-	StartBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error)
-	StopBenchmark(ctx context.Context, in *StopRequest) (resp *MemoryStatList, err error)
-	QuorumCall(ctx context.Context, in *Echo) (resp *Echo, err error)
-	AsyncQuorumCall(ctx context.Context, in *Echo) *AsyncEcho
-	SlowServer(ctx context.Context, in *Echo) (resp *Echo, err error)
+	StartServerBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse]
+	StopServerBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*Result]
+	StartBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse]
+	StopBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*MemoryStat]
+	QuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo]
+	AsyncQuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo]
+	SlowServer(ctx context.Context, in *Echo) gorums.Responses[*Echo]
 	Multicast(ctx context.Context, in *TimedMsg, opts ...gorums.CallOption)
 }
 
@@ -161,8 +129,9 @@ var _ BenchmarkClient = (*Configuration)(nil)
 // Reference imports to suppress errors if they are not otherwise used.
 var _ emptypb.Empty
 
-// Multicast is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// Multicast is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
 func (c *Configuration) Multicast(ctx context.Context, in *TimedMsg, opts ...gorums.CallOption) {
 	cd := gorums.QuorumCallData{
 		Message: in,
@@ -172,189 +141,93 @@ func (c *Configuration) Multicast(ctx context.Context, in *TimedMsg, opts ...gor
 	c.RawConfiguration.Multicast(ctx, cd, opts...)
 }
 
-// QuorumSpec is the interface of quorum functions for Benchmark.
-type QuorumSpec interface {
-	gorums.ConfigOption
+// StartServerBenchmark is a quorum call invoked on each node in configuration c,
 
-	// StartServerBenchmarkQF is the quorum function for the StartServerBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StartServerBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StartRequest'.
-	StartServerBenchmarkQF(in *StartRequest, replies map[uint32]*StartResponse) (*StartResponse, bool)
+// with the same argument in, and returns the responses as an iterator.
+func (c *Configuration) StartServerBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "benchmark.Benchmark.StartServerBenchmark",
+		ServerStream: false,
+	}
 
-	// StopServerBenchmarkQF is the quorum function for the StopServerBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StopServerBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StopRequest'.
-	StopServerBenchmarkQF(in *StopRequest, replies map[uint32]*Result) (*Result, bool)
-
-	// StartBenchmarkQF is the quorum function for the StartBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StartBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StartRequest'.
-	StartBenchmarkQF(in *StartRequest, replies map[uint32]*StartResponse) (*StartResponse, bool)
-
-	// StopBenchmarkQF is the quorum function for the StopBenchmark
-	// quorum call method. The in parameter is the request object
-	// supplied to the StopBenchmark method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *StopRequest'.
-	StopBenchmarkQF(in *StopRequest, replies map[uint32]*MemoryStat) (*MemoryStatList, bool)
-
-	// QuorumCallQF is the quorum function for the QuorumCall
-	// quorum call method. The in parameter is the request object
-	// supplied to the QuorumCall method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Echo'.
-	QuorumCallQF(in *Echo, replies map[uint32]*Echo) (*Echo, bool)
-
-	// AsyncQuorumCallQF is the quorum function for the AsyncQuorumCall
-	// asynchronous quorum call method. The in parameter is the request object
-	// supplied to the AsyncQuorumCall method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Echo'.
-	AsyncQuorumCallQF(in *Echo, replies map[uint32]*Echo) (*Echo, bool)
-
-	// SlowServerQF is the quorum function for the SlowServer
-	// quorum call method. The in parameter is the request object
-	// supplied to the SlowServer method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *Echo'.
-	SlowServerQF(in *Echo, replies map[uint32]*Echo) (*Echo, bool)
+	return gorums.QuorumCall[*StartResponse](ctx, c.RawConfiguration, cd)
 }
 
-// StartServerBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *Configuration) StartServerBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error) {
+// StopServerBenchmark is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *Configuration) StopServerBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*Result] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StartServerBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*StartResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*StartResponse)
-		}
-		return c.qspec.StartServerBenchmarkQF(req.(*StartRequest), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.StopServerBenchmark",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*StartResponse), err
+	return gorums.QuorumCall[*Result](ctx, c.RawConfiguration, cd)
 }
 
-// StopServerBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *Configuration) StopServerBenchmark(ctx context.Context, in *StopRequest) (resp *Result, err error) {
+// StartBenchmark is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *Configuration) StartBenchmark(ctx context.Context, in *StartRequest) gorums.Responses[*StartResponse] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StopServerBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Result, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Result)
-		}
-		return c.qspec.StopServerBenchmarkQF(req.(*StopRequest), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.StartBenchmark",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Result), err
+	return gorums.QuorumCall[*StartResponse](ctx, c.RawConfiguration, cd)
 }
 
-// StartBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *Configuration) StartBenchmark(ctx context.Context, in *StartRequest) (resp *StartResponse, err error) {
+// StopBenchmark is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *Configuration) StopBenchmark(ctx context.Context, in *StopRequest) gorums.Responses[*MemoryStat] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StartBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*StartResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*StartResponse)
-		}
-		return c.qspec.StartBenchmarkQF(req.(*StartRequest), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.StopBenchmark",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*StartResponse), err
-}
-
-// StopBenchmark is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *Configuration) StopBenchmark(ctx context.Context, in *StopRequest) (resp *MemoryStatList, err error) {
-	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.StopBenchmark",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*MemoryStat, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*MemoryStat)
-		}
-		return c.qspec.StopBenchmarkQF(req.(*StopRequest), r)
-	}
-
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*MemoryStatList), err
+	return gorums.QuorumCall[*MemoryStat](ctx, c.RawConfiguration, cd)
 }
 
 // benchmarks
-func (c *Configuration) QuorumCall(ctx context.Context, in *Echo) (resp *Echo, err error) {
+func (c *Configuration) QuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.QuorumCall",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Echo, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Echo)
-		}
-		return c.qspec.QuorumCallQF(req.(*Echo), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.QuorumCall",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Echo), err
+	return gorums.QuorumCall[*Echo](ctx, c.RawConfiguration, cd)
 }
 
-// SlowServer is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
-func (c *Configuration) SlowServer(ctx context.Context, in *Echo) (resp *Echo, err error) {
+// AsyncQuorumCall is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *Configuration) AsyncQuorumCall(ctx context.Context, in *Echo) gorums.Responses[*Echo] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "benchmark.Benchmark.SlowServer",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Echo, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Echo)
-		}
-		return c.qspec.SlowServerQF(req.(*Echo), r)
+		Message:      in,
+		Method:       "benchmark.Benchmark.AsyncQuorumCall",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
+	return gorums.QuorumCall[*Echo](ctx, c.RawConfiguration, cd)
+}
+
+// SlowServer is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *Configuration) SlowServer(ctx context.Context, in *Echo) gorums.Responses[*Echo] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "benchmark.Benchmark.SlowServer",
+		ServerStream: false,
 	}
-	return res.(*Echo), err
+
+	return gorums.QuorumCall[*Echo](ctx, c.RawConfiguration, cd)
 }
 
 // Benchmark is the server-side API for the Benchmark Service
@@ -417,43 +290,4 @@ func RegisterBenchmarkServer(srv *gorums.Server, impl BenchmarkServer) {
 		defer ctx.Release()
 		impl.Multicast(ctx, req)
 	})
-}
-
-type internalEcho struct {
-	nid   uint32
-	reply *Echo
-	err   error
-}
-
-type internalMemoryStat struct {
-	nid   uint32
-	reply *MemoryStat
-	err   error
-}
-
-type internalResult struct {
-	nid   uint32
-	reply *Result
-	err   error
-}
-
-type internalStartResponse struct {
-	nid   uint32
-	reply *StartResponse
-	err   error
-}
-
-// AsyncEcho is a async object for processing replies.
-type AsyncEcho struct {
-	*gorums.Async
-}
-
-// Get returns the reply and any error associated with the called method.
-// The method blocks until a reply or error is available.
-func (f *AsyncEcho) Get() (*Echo, error) {
-	resp, err := f.Async.Get()
-	if err != nil {
-		return nil, err
-	}
-	return resp.(*Echo), err
 }
