@@ -1,109 +1,73 @@
 package benchmark
 
-// QSpec is the quorum specification object for the benchmark
-type QSpec struct {
-	CfgSize int
-	QSize   int
-}
+import (
+	fmt "fmt"
 
-func (qspec *QSpec) qf(replies map[uint32]*Echo) (*Echo, bool) {
-	if len(replies) < qspec.QSize {
-		return nil, false
-	}
-	var resp *Echo
-	for _, r := range replies {
-		resp = r
-		break
-	}
-	return resp, true
-}
+	gorums "github.com/relab/gorums"
+	"google.golang.org/protobuf/proto"
+)
 
-// StartServerBenchmarkQF is the quorum function for the StartServerBenchmark quorumcall.
+// stopServerBenchmarkQF is the quorum function for the StopServerBenchmark quorumcall.
 // It requires a response from all nodes.
-func (qspec *QSpec) StartServerBenchmarkQF(_ *StartRequest, replies map[uint32]*StartResponse) (*StartResponse, bool) {
-	if len(replies) < qspec.CfgSize {
-		return nil, false
-	}
-	// return any response
-	var resp *StartResponse
-	for _, r := range replies {
-		resp = r
-		break
-	}
-	return resp, true
-}
-
-// StopServerBenchmarkQF is the quorum function for the StopServerBenchmark quorumcall.
-// It requires a response from all nodes.
-func (qspec *QSpec) StopServerBenchmarkQF(_ *StopRequest, replies map[uint32]*Result) (*Result, bool) {
-	if len(replies) < qspec.CfgSize {
-		return nil, false
-	}
+func stopServerBenchmarkQF(replies gorums.Responses[*Result], cfgSize int) (*Result, error) {
 	// combine results, calculating averages and pooled variance
-	resp := &Result{}
-	for _, reply := range replies {
-		if resp.GetName() != "" {
-			resp.SetName(reply.GetName())
+	resp := &Result_builder{}
+	replyCount := 0
+	for reply := range replies.IgnoreErrors() {
+		replyCount++
+		msg := reply.Msg
+		if resp.Name != "" {
+			resp.Name = msg.GetName()
 		}
-		resp.SetTotalOps(resp.GetTotalOps() + reply.GetTotalOps())
-		resp.SetTotalTime(resp.GetTotalTime() + reply.GetTotalTime())
-		resp.SetThroughput(resp.GetThroughput() + reply.GetThroughput())
-		resp.SetLatencyAvg(resp.GetLatencyAvg() + reply.GetLatencyAvg()*float64(reply.GetTotalOps()))
-		resp.SetServerStats(append(resp.GetServerStats(), MemoryStat_builder{
-			Allocs: reply.GetAllocsPerOp() * resp.GetTotalOps(),
-			Memory: reply.GetMemPerOp() * resp.GetTotalOps(),
-		}.Build()))
+		resp.TotalOps += msg.GetTotalOps()
+		resp.TotalTime += msg.GetTotalTime()
+		resp.Throughput += msg.GetThroughput()
+		resp.LatencyAvg += msg.GetLatencyAvg() * float64(msg.GetTotalOps())
+		resp.LatencyVar += float64(msg.GetTotalOps()-1) * msg.GetLatencyVar()
+		resp.ServerStats = append(resp.ServerStats, MemoryStat_builder{
+			Allocs: msg.GetAllocsPerOp() * resp.TotalOps,
+			Memory: msg.GetMemPerOp() * resp.TotalOps,
+		}.Build())
 	}
-	resp.SetLatencyAvg(resp.GetLatencyAvg() / float64(resp.GetTotalOps()))
-	for _, reply := range replies {
-		resp.SetLatencyVar(resp.GetLatencyVar() + float64(reply.GetTotalOps()-1)*reply.GetLatencyVar())
+
+	if replyCount < cfgSize {
+		return nil, fmt.Errorf("missing benchmark results got %d want %d", replyCount, cfgSize)
 	}
-	resp.SetLatencyVar(resp.GetLatencyVar() / (float64(resp.GetTotalOps()) - float64(len(replies))))
-	resp.SetTotalOps(resp.GetTotalOps() / uint64(len(replies)))
-	resp.SetTotalTime(resp.GetTotalTime() / int64(len(replies)))
-	resp.SetThroughput(resp.GetThroughput() / float64(len(replies)))
-	return resp, true
+
+	resp.LatencyAvg /= float64(resp.TotalOps)
+	resp.LatencyVar /= float64(resp.TotalOps) - float64(replyCount)
+	resp.TotalOps /= uint64(replyCount)
+	resp.TotalTime /= int64(replyCount)
+	resp.Throughput /= float64(replyCount)
+	return resp.Build(), nil
 }
 
-// StartBenchmarkQF is the quorum function for the StartBenchmark quorumcall.
+// stopBenchmarkQF is the quorum function for the StopBenchmark quorumcall.
 // It requires a response from all nodes.
-func (qspec *QSpec) StartBenchmarkQF(_ *StartRequest, replies map[uint32]*StartResponse) (*StartResponse, bool) {
-	if len(replies) < qspec.CfgSize {
-		return nil, false
+func stopBenchmarkQF(replies gorums.Responses[*MemoryStat], cfgSize int) ([]*MemoryStat, error) {
+	replyList := make([]*MemoryStat, 0)
+	replyCount := 0
+	for reply := range replies.IgnoreErrors() {
+		replyCount++
+		replyList = append(replyList, reply.Msg)
+		if len(replyList) < cfgSize {
+			continue
+		}
+		return replyList, nil
 	}
-	// return any response
-	var resp *StartResponse
-	for _, r := range replies {
-		resp = r
-		break
-	}
-	return resp, true
+	return nil, fmt.Errorf("quorum not found, got %d responses want %d", replyCount, cfgSize)
 }
 
-// StopBenchmarkQF is the quorum function for the StopBenchmark quorumcall.
-// It requires a response from all nodes.
-func (qspec *QSpec) StopBenchmarkQF(_ *StopRequest, replies map[uint32]*MemoryStat) (*MemoryStatList, bool) {
-	if len(replies) < qspec.CfgSize {
-		return nil, false
+// qf is a generic quorum function which does not return anything.
+// It requires a response from a specified quorum amount of nodes.
+func qf[responseType proto.Message](replies gorums.Responses[responseType], quorum int) error {
+	replyCount := 0
+	for range replies.IgnoreErrors() {
+		replyCount++
+		if replyCount < quorum {
+			continue
+		}
+		return nil
 	}
-	replyList := make([]*MemoryStat, 0, len(replies))
-	for _, v := range replies {
-		replyList = append(replyList, v)
-	}
-	return MemoryStatList_builder{MemoryStats: replyList}.Build(), true
-}
-
-// QuorumCallQF is the quorum function for the QuorumCall quorumcall
-func (qspec *QSpec) QuorumCallQF(_ *Echo, replies map[uint32]*Echo) (*Echo, bool) {
-	return qspec.qf(replies)
-}
-
-// AsyncQuorumCallQF is the quorum function for the AsyncQuorumCall quorumcall
-func (qspec *QSpec) AsyncQuorumCallQF(_ *Echo, replies map[uint32]*Echo) (*Echo, bool) {
-	return qspec.qf(replies)
-}
-
-// SlowServerQF is the quorum function for the SlowServer quorumcall
-func (qspec *QSpec) SlowServerQF(_ *Echo, replies map[uint32]*Echo) (*Echo, bool) {
-	return qspec.qf(replies)
+	return fmt.Errorf("quorum not found, got %d responses want %d", replyCount, quorum)
 }
