@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+125b0cfb
+// 	protoc            v6.30.2
 // source: zorums.proto
 
 package dev
@@ -9,15 +9,15 @@ package dev
 import (
 	context "context"
 	gorums "github.com/relab/gorums"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	proto "google.golang.org/protobuf/proto"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // Multicast plain. Response type is not needed here.
@@ -37,7 +37,7 @@ func (c *Configuration) MulticastPerNodeArg(ctx context.Context, in *Request, f 
 		Method:  "dev.ZorumsService.MulticastPerNodeArg",
 	}
 
-	cd.PerNodeArgFn = func(req protoreflect.ProtoMessage, nid uint32) protoreflect.ProtoMessage {
+	cd.PerNodeArgFn = func(req proto.Message, nid uint32) proto.Message {
 		return f(req.(*Request), nid)
 	}
 
