@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+125b0cfb
+// 	protoc            v6.30.2
 // source: zorums.proto
 
 package dev
@@ -9,145 +9,107 @@ package dev
 import (
 	context "context"
 	gorums "github.com/relab/gorums"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	proto "google.golang.org/protobuf/proto"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // QuorumCall plain.
-func (c *Configuration) QuorumCall(ctx context.Context, in *Request) (resp *Response, err error) {
+func (c *Configuration) QuorumCall(ctx context.Context, in *Request) gorums.Responses[*Response] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "dev.ZorumsService.QuorumCall",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.QuorumCallQF(req.(*Request), r)
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCall",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Response), err
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
 }
 
 // QuorumCall with per_node_arg option.
-func (c *Configuration) QuorumCallPerNodeArg(ctx context.Context, in *Request, f func(*Request, uint32) *Request) (resp *Response, err error) {
+func (c *Configuration) QuorumCallPerNodeArg(ctx context.Context, in *Request, f func(*Request, uint32) *Request) gorums.Responses[*Response] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "dev.ZorumsService.QuorumCallPerNodeArg",
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCallPerNodeArg",
+		ServerStream: false,
 	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.QuorumCallPerNodeArgQF(req.(*Request), r)
-	}
-	cd.PerNodeArgFn = func(req protoreflect.ProtoMessage, nid uint32) protoreflect.ProtoMessage {
+	cd.PerNodeArgFn = func(req proto.Message, nid uint32) proto.Message {
 		return f(req.(*Request), nid)
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Response), err
-}
-
-// QuorumCall with custom_return_type option.
-func (c *Configuration) QuorumCallCustomReturnType(ctx context.Context, in *Request) (resp *MyResponse, err error) {
-	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "dev.ZorumsService.QuorumCallCustomReturnType",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.QuorumCallCustomReturnTypeQF(req.(*Request), r)
-	}
-
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*MyResponse), err
-}
-
-// QuorumCallCombo with all supported options.
-func (c *Configuration) QuorumCallCombo(ctx context.Context, in *Request, f func(*Request, uint32) *Request) (resp *MyResponse, err error) {
-	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "dev.ZorumsService.QuorumCallCombo",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.QuorumCallComboQF(req.(*Request), r)
-	}
-	cd.PerNodeArgFn = func(req protoreflect.ProtoMessage, nid uint32) protoreflect.ProtoMessage {
-		return f(req.(*Request), nid)
-	}
-
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*MyResponse), err
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
 }
 
 // QuorumCallEmpty for testing imported message type.
-func (c *Configuration) QuorumCallEmpty(ctx context.Context, in *emptypb.Empty) (resp *Response, err error) {
+func (c *Configuration) QuorumCallEmpty(ctx context.Context, in *emptypb.Empty) gorums.Responses[*Response] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "dev.ZorumsService.QuorumCallEmpty",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*Response, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*Response)
-		}
-		return c.qspec.QuorumCallEmptyQF(req.(*emptypb.Empty), r)
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCallEmpty",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*Response), err
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
 }
 
 // QuorumCallEmpty2 for testing imported message type.
-func (c *Configuration) QuorumCallEmpty2(ctx context.Context, in *Request) (resp *emptypb.Empty, err error) {
+func (c *Configuration) QuorumCallEmpty2(ctx context.Context, in *Request) gorums.Responses[*emptypb.Empty] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "dev.ZorumsService.QuorumCallEmpty2",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*emptypb.Empty, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*emptypb.Empty)
-		}
-		return c.qspec.QuorumCallEmpty2QF(req.(*Request), r)
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCallEmpty2",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
+	return gorums.QuorumCall[*emptypb.Empty](ctx, c.RawConfiguration, cd)
+}
+
+// QuorumCall plain.
+func (c *Configuration) QuorumCallStream(ctx context.Context, in *Request) gorums.Responses[*Response] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCallStream",
+		ServerStream: true,
 	}
-	return res.(*emptypb.Empty), err
+
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
+}
+
+// QuorumCall with per_node_arg option.
+func (c *Configuration) QuorumCallStreamPerNodeArg(ctx context.Context, in *Request, f func(*Request, uint32) *Request) gorums.Responses[*Response] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCallStreamPerNodeArg",
+		ServerStream: true,
+	}
+	cd.PerNodeArgFn = func(req proto.Message, nid uint32) proto.Message {
+		return f(req.(*Request), nid)
+	}
+
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
+}
+
+// QuorumCallEmpty for testing imported message type.
+func (c *Configuration) QuorumCallStreamEmpty(ctx context.Context, in *emptypb.Empty) gorums.Responses[*Response] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCallStreamEmpty",
+		ServerStream: true,
+	}
+
+	return gorums.QuorumCall[*Response](ctx, c.RawConfiguration, cd)
+}
+
+// QuorumCallEmpty2 for testing imported message type.
+func (c *Configuration) QuorumCallStreamEmpty2(ctx context.Context, in *Request) gorums.Responses[*emptypb.Empty] {
+	cd := gorums.QuorumCallData{
+		Message:      in,
+		Method:       "dev.ZorumsService.QuorumCallStreamEmpty2",
+		ServerStream: true,
+	}
+
+	return gorums.QuorumCall[*emptypb.Empty](ctx, c.RawConfiguration, cd)
 }
