@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+125b0cfb
+// 	protoc            v6.30.2
 // source: zorums.proto
 
 package dev
@@ -13,9 +13,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // GRPCCall plain gRPC call; testing that Gorums can ignore these, but that
