@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+125b0cfb
+// 	protoc            v6.30.2
 // source: zorums.proto
 
 package dev
@@ -15,9 +15,9 @@ import (
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // ZorumsService is the server-side API for the ZorumsService Service
@@ -25,34 +25,17 @@ type ZorumsServiceServer interface {
 	GRPCCall(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 	QuorumCall(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 	QuorumCallPerNodeArg(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	QuorumCallCustomReturnType(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	QuorumCallCombo(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
 	QuorumCallEmpty(ctx gorums.ServerCtx, request *emptypb.Empty) (response *Response, err error)
 	QuorumCallEmpty2(ctx gorums.ServerCtx, request *Request) (response *emptypb.Empty, err error)
+	QuorumCallStream(ctx gorums.ServerCtx, request *Request, send func(response *Response) error) error
+	QuorumCallStreamPerNodeArg(ctx gorums.ServerCtx, request *Request, send func(response *Response) error) error
+	QuorumCallStreamEmpty(ctx gorums.ServerCtx, request *emptypb.Empty, send func(response *Response) error) error
+	QuorumCallStreamEmpty2(ctx gorums.ServerCtx, request *Request, send func(response *emptypb.Empty) error) error
 	Multicast(ctx gorums.ServerCtx, request *Request)
 	MulticastPerNodeArg(ctx gorums.ServerCtx, request *Request)
 	Multicast2(ctx gorums.ServerCtx, request *Request)
 	Multicast3(ctx gorums.ServerCtx, request *Request)
 	Multicast4(ctx gorums.ServerCtx, request *emptypb.Empty)
-	QuorumCallAsync(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	QuorumCallAsyncPerNodeArg(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	QuorumCallAsyncCustomReturnType(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	QuorumCallAsyncCombo(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	QuorumCallAsync2(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	QuorumCallAsyncEmpty(ctx gorums.ServerCtx, request *Request) (response *emptypb.Empty, err error)
-	QuorumCallAsyncEmpty2(ctx gorums.ServerCtx, request *emptypb.Empty) (response *Response, err error)
-	Correctable(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	CorrectablePerNodeArg(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	CorrectableCustomReturnType(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	CorrectableCombo(ctx gorums.ServerCtx, request *Request) (response *Response, err error)
-	CorrectableEmpty(ctx gorums.ServerCtx, request *Request) (response *emptypb.Empty, err error)
-	CorrectableEmpty2(ctx gorums.ServerCtx, request *emptypb.Empty) (response *Response, err error)
-	CorrectableStream(ctx gorums.ServerCtx, request *Request, send func(response *Response) error) error
-	CorrectableStreamPerNodeArg(ctx gorums.ServerCtx, request *Request, send func(response *Response) error) error
-	CorrectableStreamCustomReturnType(ctx gorums.ServerCtx, request *Request, send func(response *Response) error) error
-	CorrectableStreamCombo(ctx gorums.ServerCtx, request *Request, send func(response *Response) error) error
-	CorrectableStreamEmpty(ctx gorums.ServerCtx, request *Request, send func(response *emptypb.Empty) error) error
-	CorrectableStreamEmpty2(ctx gorums.ServerCtx, request *emptypb.Empty, send func(response *Response) error) error
 	Unicast(ctx gorums.ServerCtx, request *Request)
 	Unicast2(ctx gorums.ServerCtx, request *Request)
 }
@@ -76,18 +59,6 @@ func RegisterZorumsServiceServer(srv *gorums.Server, impl ZorumsServiceServer) {
 		resp, err := impl.QuorumCallPerNodeArg(ctx, req)
 		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
 	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallCustomReturnType", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallCustomReturnType(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallCombo", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallCombo(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
 	srv.RegisterHandler("dev.ZorumsService.QuorumCallEmpty", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
 		req := in.Message.(*emptypb.Empty)
 		defer ctx.Release()
@@ -99,6 +70,54 @@ func RegisterZorumsServiceServer(srv *gorums.Server, impl ZorumsServiceServer) {
 		defer ctx.Release()
 		resp, err := impl.QuorumCallEmpty2(ctx, req)
 		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
+	})
+	srv.RegisterHandler("dev.ZorumsService.QuorumCallStream", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+		req := in.Message.(*Request)
+		defer ctx.Release()
+		err := impl.QuorumCallStream(ctx, req, func(resp *Response) error {
+			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
+			md := proto.Clone(in.Metadata)
+			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
+		})
+		if err != nil {
+			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
+		}
+	})
+	srv.RegisterHandler("dev.ZorumsService.QuorumCallStreamPerNodeArg", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+		req := in.Message.(*Request)
+		defer ctx.Release()
+		err := impl.QuorumCallStreamPerNodeArg(ctx, req, func(resp *Response) error {
+			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
+			md := proto.Clone(in.Metadata)
+			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
+		})
+		if err != nil {
+			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
+		}
+	})
+	srv.RegisterHandler("dev.ZorumsService.QuorumCallStreamEmpty", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+		req := in.Message.(*emptypb.Empty)
+		defer ctx.Release()
+		err := impl.QuorumCallStreamEmpty(ctx, req, func(resp *Response) error {
+			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
+			md := proto.Clone(in.Metadata)
+			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
+		})
+		if err != nil {
+			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
+		}
+	})
+	srv.RegisterHandler("dev.ZorumsService.QuorumCallStreamEmpty2", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+		req := in.Message.(*Request)
+		defer ctx.Release()
+		err := impl.QuorumCallStreamEmpty2(ctx, req, func(resp *emptypb.Empty) error {
+			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
+			md := proto.Clone(in.Metadata)
+			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
+		})
+		if err != nil {
+			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
+		}
 	})
 	srv.RegisterHandler("dev.ZorumsService.Multicast", func(ctx gorums.ServerCtx, in *gorums.Message, _ chan<- *gorums.Message) {
 		req := in.Message.(*Request)
@@ -124,156 +143,6 @@ func RegisterZorumsServiceServer(srv *gorums.Server, impl ZorumsServiceServer) {
 		req := in.Message.(*emptypb.Empty)
 		defer ctx.Release()
 		impl.Multicast4(ctx, req)
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallAsync", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallAsync(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallAsyncPerNodeArg", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallAsyncPerNodeArg(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallAsyncCustomReturnType", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallAsyncCustomReturnType(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallAsyncCombo", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallAsyncCombo(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallAsync2", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallAsync2(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallAsyncEmpty", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallAsyncEmpty(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.QuorumCallAsyncEmpty2", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*emptypb.Empty)
-		defer ctx.Release()
-		resp, err := impl.QuorumCallAsyncEmpty2(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.Correctable", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.Correctable(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectablePerNodeArg", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.CorrectablePerNodeArg(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableCustomReturnType", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.CorrectableCustomReturnType(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableCombo", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.CorrectableCombo(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableEmpty", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		resp, err := impl.CorrectableEmpty(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableEmpty2", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*emptypb.Empty)
-		defer ctx.Release()
-		resp, err := impl.CorrectableEmpty2(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableStream", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		err := impl.CorrectableStream(ctx, req, func(resp *Response) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := proto.Clone(in.Metadata)
-			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
-		})
-		if err != nil {
-			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
-		}
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableStreamPerNodeArg", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		err := impl.CorrectableStreamPerNodeArg(ctx, req, func(resp *Response) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := proto.Clone(in.Metadata)
-			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
-		})
-		if err != nil {
-			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
-		}
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableStreamCustomReturnType", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		err := impl.CorrectableStreamCustomReturnType(ctx, req, func(resp *Response) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := proto.Clone(in.Metadata)
-			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
-		})
-		if err != nil {
-			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
-		}
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableStreamCombo", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		err := impl.CorrectableStreamCombo(ctx, req, func(resp *Response) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := proto.Clone(in.Metadata)
-			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
-		})
-		if err != nil {
-			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
-		}
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableStreamEmpty", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*Request)
-		defer ctx.Release()
-		err := impl.CorrectableStreamEmpty(ctx, req, func(resp *emptypb.Empty) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := proto.Clone(in.Metadata)
-			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
-		})
-		if err != nil {
-			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
-		}
-	})
-	srv.RegisterHandler("dev.ZorumsService.CorrectableStreamEmpty2", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
-		req := in.Message.(*emptypb.Empty)
-		defer ctx.Release()
-		err := impl.CorrectableStreamEmpty2(ctx, req, func(resp *Response) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := proto.Clone(in.Metadata)
-			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
-		})
-		if err != nil {
-			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
-		}
 	})
 	srv.RegisterHandler("dev.ZorumsService.Unicast", func(ctx gorums.ServerCtx, in *gorums.Message, _ chan<- *gorums.Message) {
 		req := in.Message.(*Request)
