@@ -8,7 +8,6 @@ import (
 	"strings"
 
 	"github.com/relab/gorums"
-	"github.com/relab/gorums/internal/correctable"
 	"github.com/relab/gorums/internal/version"
 	"google.golang.org/protobuf/compiler/protogen"
 	"google.golang.org/protobuf/proto"
@@ -148,34 +147,6 @@ func genGorumsMethods(gorumsType string, data servicesData, callTypeInfo *callTy
 	}
 }
 
-// callType returns call type information for the given method.
-// If the given method has specified a Gorums method option that
-// correspond to a call type, this call type is returned. Further,
-// if the call type has a sub call type, then this is returned instead.
-func callType(method *protogen.Method) *callTypeInfo {
-	for _, callTypeInfo := range callTypeOptions(method) {
-		callType := callTypeInfo.deriveCallType(method)
-		if callType.check(method) {
-			return callType
-		}
-	}
-	panic(fmt.Sprintf("unknown call type for method %s\n", method.GoName))
-}
-
-// callTypeOptions returns all Gorums call types for the given method.
-func callTypeOptions(method *protogen.Method) []*callTypeInfo {
-	methExt := protoimpl.X.MessageOf(method.Desc.Options()).Interface()
-	var options []*callTypeInfo
-	for _, callType := range gorumsCallTypesInfo {
-		if callType.extInfo != nil {
-			if proto.HasExtension(methExt, callType.extInfo) {
-				options = append(options, callType)
-			}
-		}
-	}
-	return options
-}
-
 // hasGorumsMethods returns true if one of the methods in the set of services
 // has a Gorums call type.
 func hasGorumsMethods(services []*protogen.Service) bool {
@@ -200,7 +171,6 @@ type callTypeInfo struct {
 	extInfo        *protoimpl.ExtensionInfo
 	docName        string
 	template       string
-	outPrefix      string
 	chkFn          func(m *protogen.Method) bool
 	nestedCallType map[string]*callTypeInfo
 }
@@ -239,8 +209,6 @@ func callTypeName(ext *protoimpl.ExtensionInfo) string {
 // The entries in this map is used to generate dev/zorums_{type}.pb.go
 // files for the different keys.
 var gorumsCallTypesInfo = map[string]*callTypeInfo{
-	"qspec":  {template: qspecInterface},
-	"types":  {template: dataTypes},
 	"server": {template: server},
 	"client": {template: client},
 
@@ -257,42 +225,7 @@ var gorumsCallTypesInfo = map[string]*callTypeInfo{
 		docName:  "quorum",
 		template: quorumCall,
 		chkFn: func(m *protogen.Method) bool {
-			return hasMethodOption(m, gorums.E_Quorumcall) && !hasMethodOption(m, gorums.E_Async)
-		},
-	},
-	callTypeName(gorums.E_Async): {
-		extInfo:   gorums.E_Async,
-		docName:   "asynchronous quorum",
-		template:  asyncCall,
-		outPrefix: "Async",
-		chkFn: func(m *protogen.Method) bool {
-			return hasAllMethodOption(m, gorums.E_Quorumcall, gorums.E_Async)
-		},
-	},
-	callTypeName(gorums.E_Correctable): {
-		extInfo: gorums.E_Correctable,
-		chkFn: func(m *protogen.Method) bool {
-			return hasMethodOption(m, gorums.E_Correctable)
-		},
-		nestedCallType: map[string]*callTypeInfo{
-			callTypeName(correctable.E_Correctable): {
-				extInfo:   correctable.E_Correctable,
-				docName:   "correctable quorum",
-				template:  correctableCall,
-				outPrefix: "Correctable",
-				chkFn: func(m *protogen.Method) bool {
-					return hasMethodOption(m, gorums.E_Correctable) && !m.Desc.IsStreamingServer()
-				},
-			},
-			callTypeName(correctable.E_CorrectableStream): {
-				extInfo:   correctable.E_CorrectableStream,
-				docName:   "correctable stream quorum",
-				template:  correctableCall,
-				outPrefix: "CorrectableStream",
-				chkFn: func(m *protogen.Method) bool {
-					return hasMethodOption(m, gorums.E_Correctable) && m.Desc.IsStreamingServer()
-				},
-			},
+			return hasMethodOption(m, gorums.E_Quorumcall)
 		},
 	},
 	callTypeName(gorums.E_Multicast): {
@@ -317,26 +250,8 @@ var gorumsCallTypesInfo = map[string]*callTypeInfo{
 // These are considered mutually incompatible.
 var gorumsCallTypes = []*protoimpl.ExtensionInfo{
 	gorums.E_Quorumcall,
-	gorums.E_Async,
-	gorums.E_Correctable,
 	gorums.E_Multicast,
 	gorums.E_Unicast,
-}
-
-// callTypesWithInternal should list all available call types that
-// has a quorum function and hence need an internal type that wraps
-// the return type with additional information.
-var callTypesWithInternal = []*protoimpl.ExtensionInfo{
-	gorums.E_Quorumcall,
-	gorums.E_Async,
-	gorums.E_Correctable,
-}
-
-// callTypesWithPromiseObject lists all call types that returns
-// a promise (async or correctable) object.
-var callTypesWithPromiseObject = []*protoimpl.ExtensionInfo{
-	gorums.E_Async,
-	gorums.E_Correctable,
 }
 
 // hasGorumsCallType returns true if the given method has specified
@@ -360,32 +275,11 @@ func hasMethodOption(method *protogen.Method, methodOptions ...*protoimpl.Extens
 	return false
 }
 
-// hasAllMethodOption returns true if the method has all of the given method options
-func hasAllMethodOption(method *protogen.Method, methodOptions ...*protoimpl.ExtensionInfo) bool {
-	ext := protoimpl.X.MessageOf(method.Desc.Options()).Interface()
-	for _, callType := range methodOptions {
-		if !proto.HasExtension(ext, callType) {
-			return false
-		}
-	}
-	return true
-}
-
 // validateOptions returns an error if the extension options
 // for the provided method are incompatible.
 func validateOptions(method *protogen.Method) error {
-	switch {
-	case hasMethodOption(method, gorums.E_Async) && !hasMethodOption(method, gorums.E_Quorumcall):
-		return optionErrorf("is required for async methods", method, gorums.E_Quorumcall)
-
-	case !hasMethodOption(method, gorums.E_Multicast) && method.Desc.IsStreamingClient():
+	if !hasMethodOption(method, gorums.E_Multicast) && method.Desc.IsStreamingClient() {
 		return optionErrorf("is required for client-server stream methods", method, gorums.E_Multicast)
-
-	case !hasMethodOption(method, gorums.E_Correctable) && method.Desc.IsStreamingServer():
-		return optionErrorf("is required for server-client stream methods", method, gorums.E_Correctable)
-
-	case hasMethodOption(method, gorums.E_Correctable) && method.Desc.IsStreamingClient():
-		return optionErrorf("is only valid for server-client stream methods", method, gorums.E_Correctable)
 	}
 	return nil
 }
