@@ -22,12 +22,10 @@ var clientConfigurationInterface = `
 			{{- $method := .GoName}}
 			{{- if isOneway .}}
 				{{$method}}(ctx {{$context}}, in *{{in $genFile .}} {{perNodeFnType $genFile . ", f"}}, opts ...{{$callOpt}})
-			{{- else if or (isCorrectable .) (isAsync .)}}
-				{{- $customOut := outType . (customOut $genFile .)}}
-				{{$method}}(ctx {{$context}}, in *{{in $genFile .}} {{perNodeFnType $genFile . ", f"}}) *{{$customOut}}
 			{{- else}}
-				{{- $customOut := customOut $genFile .}}
-				{{$method}}(ctx {{$context}}, in *{{in $genFile .}} {{perNodeFnType $genFile . ", f"}}) (resp *{{$customOut}}, err error)
+				{{- $out := out $genFile .}}
+				{{- $iterator := use "gorums.Responses" $genFile}}
+				{{$method}}(ctx {{$context}}, in *{{in $genFile .}} {{perNodeFnType $genFile . ", f"}}) {{$iterator}}[*{{$out}}]
 			{{- end}}
 		{{- end}}
 	}
@@ -48,8 +46,8 @@ var clientNodeInterface = `
 			{{- if isOneway .}}
 				{{$method}}(ctx {{$context}}, in *{{in $genFile .}} {{perNodeFnType $genFile . ", f"}}, opts ...{{$callOpt}})
 			{{- else}}
-				{{- $customOut := customOut $genFile .}}
-				{{$method}}(ctx {{$context}}, in *{{in $genFile .}} {{perNodeFnType $genFile . ", f"}}) (resp *{{$customOut}}, err error)
+				{{- $out := out $genFile .}}
+				{{$method}}(ctx {{$context}}, in *{{in $genFile .}} {{perNodeFnType $genFile . ", f"}}) (resp *{{$out}}, err error)
 			{{- end}}
 		{{- end}}
 	}
