@@ -11,13 +11,13 @@ var serverInterface = `
 // {{$service}} is the server-side API for the {{$service}} Service
 type {{$service}}Server interface {
 	{{- range .Methods}}
-	{{- if isOneway .}}
-	{{.GoName}}(ctx {{$context}}, request *{{in $genFile .}})
-	{{- else if correctableStream .}}
-	{{.GoName}}(ctx {{$context}}, request *{{in $genFile .}}, send func(response *{{out $genFile .}}) error) error
-	{{- else}}
-	{{.GoName}}(ctx {{$context}}, request *{{in $genFile .}}) (response *{{out $genFile .}}, err error)
-	{{- end}}
+		{{- if isOneway .}}
+			{{.GoName}}(ctx {{$context}}, request *{{in $genFile .}})
+		{{- else if isStream .}}
+			{{.GoName}}(ctx {{$context}}, request *{{in $genFile .}}, send func(response *{{out $genFile .}}) error) error
+		{{- else}}
+			{{.GoName}}(ctx {{$context}}, request *{{in $genFile .}}) (response *{{out $genFile .}}, err error)
+		{{- end}}
 	{{- end}}
 }
 {{- end}}
@@ -36,19 +36,19 @@ func Register{{$service}}Server(srv *{{use "gorums.Server" $genFile}}, impl {{$s
 		req := in.Message.(*{{in $genFile .}})
 		defer ctx.Release()
 		{{- if isOneway .}}
-		impl.{{.GoName}}(ctx, req)
-		{{- else if correctableStream .}}
-		err := impl.{{.GoName}}(ctx, req, func(resp *{{out $genFile .}}) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := {{use "proto.Clone" $genFile}}(in.Metadata)
-			return {{$sendMessage}}(ctx, finished, {{$wrapMessage}}(md.(*{{use "ordering.Metadata" $genFile}}), resp, nil))
-		})
-		if err != nil {
-			{{$sendMessage}}(ctx, finished, {{$wrapMessage}}(in.Metadata, nil, err))
-		}
+			impl.{{.GoName}}(ctx, req)
+		{{- else if isStream .}}
+			err := impl.{{.GoName}}(ctx, req, func(resp *{{out $genFile .}}) error {
+				// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
+				md := {{use "proto.Clone" $genFile}}(in.Metadata)
+				return {{$sendMessage}}(ctx, finished, {{$wrapMessage}}(md.(*{{use "ordering.Metadata" $genFile}}), resp, nil))
+			})
+			if err != nil {
+				{{$sendMessage}}(ctx, finished, {{$wrapMessage}}(in.Metadata, nil, err))
+			}
 		{{- else }}
-		resp, err := impl.{{.GoName}}(ctx, req)
-		{{$sendMessage}}(ctx, finished, {{$wrapMessage}}(in.Metadata, resp, err))
+			resp, err := impl.{{.GoName}}(ctx, req)
+			{{$sendMessage}}(ctx, finished, {{$wrapMessage}}(in.Metadata, resp, err))
 		{{- end}}
 	})
 	{{- end}}
