@@ -1,98 +1,76 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+125b0cfb
+// 	protoc            v6.30.2
 // source: storage/proto/storage.proto
 
 package proto
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	encoding "google.golang.org/grpc/encoding"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
 // procedure calls may be invoked.
 type Configuration struct {
 	gorums.RawConfiguration
-	qspec QuorumSpec
-	nodes []*Node
 }
 
-func NewConfiguration(qspec QuorumSpec, cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *Configuration, err error) {
-	if qspec == nil {
-		return nil, fmt.Errorf("config: QuorumSpec cannot be nil")
-	}
-	c = &Configuration{
-		qspec: qspec,
-	}
+// NewConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *Configuration, err error) {
+	c = &Configuration{}
 	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
 	if err != nil {
 		return nil, err
 	}
-	c.nodes = make([]*Node, c.Size())
-	for i, n := range c.RawConfiguration.RawNodes {
-		c.nodes[i] = &Node{n}
-	}
 	return c, nil
 }
 
-// ConfigurationFromRaw returns a new Configuration from the given raw configuration and QuorumSpec.
+// ConfigurationFromRaw returns a new configuration from the given raw configuration.
 //
-// This function may for example be used to "clone" a configuration but install a different QuorumSpec:
-//
-//	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
-//	cfg2 := ConfigurationFromRaw(cfg1.RawConfiguration, qspec2)
-func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
-	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
-	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+//	cfg1, err := pb.NewConfiguration(nodeList, opts...)
+//	cfg2 := ConfigurationFromRaw(cfg1.RawConfig)
+func ConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*Configuration, error) {
 	newCfg := &Configuration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*Node, newCfg.Size())
-	for i, n := range rawCfg.Nodes() {
-		newCfg.nodes[i] = &Node{n}
 	}
 	return newCfg, nil
 }
 
-func (c *Configuration) SubConfiguration(qspec QuorumSpec, cfg gorums.NodeListOption) (subCfg *Configuration, err error) {
-	if qspec == nil {
-		return nil, fmt.Errorf("config: QuorumSpec cannot be nil")
-	}
-	subCfg = &Configuration{
-		qspec: qspec,
-	}
+// SubConfiguration allows for making a new Configuration from the ManagerOption list and
+// node list of another configuration,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *Configuration) SubConfiguration(cfg gorums.NodeListOption) (subCfg *Configuration, err error) {
+	subCfg = &Configuration{}
 	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
 	if err != nil {
 		return nil, err
-	}
-	subCfg.nodes = make([]*Node, subCfg.Size())
-	for i, n := range subCfg.RawConfiguration.Nodes() {
-		subCfg.nodes[i] = &Node{n}
 	}
 	return subCfg, nil
 }
 
 // Close closes a configuration created from the NewConfiguration method
 //
-// NOTE: A configuration created with ConfigurationFromRaw is closed when the original configuration is closed
+// NOTE: A configuration created with ConfigurationFromRaw or SubConfiguration closes and
+// is closed when the original configuration or any of the subconfigurations are closed.
 // If you want the configurations to be independent you need to use NewConfiguration
 func (c *Configuration) Close() error {
 	return c.RawConfiguration.Close()
@@ -103,17 +81,21 @@ func (c *Configuration) Close() error {
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *Configuration) Nodes() []*Node {
-	return c.nodes
+	nodes := make([]*Node, c.Size())
+	for i, n := range c.RawConfiguration.RawNodes {
+		nodes[i] = &Node{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
-func (c Configuration) And(d *Configuration) gorums.NodeListOption {
+func (c *Configuration) And(d *Configuration) gorums.NodeListOption {
 	return c.RawConfiguration.And(d.RawConfiguration)
 }
 
 // Except returns a NodeListOption that can be used to create a new configuration
 // from c without the nodes in rm.
-func (c Configuration) Except(rm *Configuration) gorums.NodeListOption {
+func (c *Configuration) Except(rm *Configuration) gorums.NodeListOption {
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
@@ -131,8 +113,8 @@ type Node struct {
 
 // StorageClient is the client interface for the Storage service.
 type StorageClient interface {
-	ReadQC(ctx context.Context, in *ReadRequest) (resp *ReadResponse, err error)
-	WriteQC(ctx context.Context, in *WriteRequest) (resp *WriteResponse, err error)
+	ReadQC(ctx context.Context, in *ReadRequest) gorums.Responses[*ReadResponse]
+	WriteQC(ctx context.Context, in *WriteRequest) gorums.Responses[*WriteResponse]
 	WriteMulticast(ctx context.Context, in *WriteRequest, opts ...gorums.CallOption)
 }
 
@@ -151,8 +133,9 @@ var _ StorageNodeClient = (*Node)(nil)
 // Reference imports to suppress errors if they are not otherwise used.
 var _ emptypb.Empty
 
-// WriteMulticast is a quorum call invoked on all nodes in configuration c,
-// with the same argument in, and returns a combined result.
+// WriteMulticast is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
 func (c *Configuration) WriteMulticast(ctx context.Context, in *WriteRequest, opts ...gorums.CallOption) {
 	cd := gorums.QuorumCallData{
 		Message: in,
@@ -162,67 +145,28 @@ func (c *Configuration) WriteMulticast(ctx context.Context, in *WriteRequest, op
 	c.RawConfiguration.Multicast(ctx, cd, opts...)
 }
 
-// QuorumSpec is the interface of quorum functions for Storage.
-type QuorumSpec interface {
-	gorums.ConfigOption
-
-	// ReadQCQF is the quorum function for the ReadQC
-	// quorum call method. The in parameter is the request object
-	// supplied to the ReadQC method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *ReadRequest'.
-	ReadQCQF(in *ReadRequest, replies map[uint32]*ReadResponse) (*ReadResponse, bool)
-
-	// WriteQCQF is the quorum function for the WriteQC
-	// quorum call method. The in parameter is the request object
-	// supplied to the WriteQC method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *WriteRequest'.
-	WriteQCQF(in *WriteRequest, replies map[uint32]*WriteResponse) (*WriteResponse, bool)
-}
-
 // ReadQC executes the Read Quorum Call on a configuration
 // of Nodes and returns the most recent value.
-func (c *Configuration) ReadQC(ctx context.Context, in *ReadRequest) (resp *ReadResponse, err error) {
+func (c *Configuration) ReadQC(ctx context.Context, in *ReadRequest) gorums.Responses[*ReadResponse] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "storage.Storage.ReadQC",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*ReadResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*ReadResponse)
-		}
-		return c.qspec.ReadQCQF(req.(*ReadRequest), r)
+		Message:      in,
+		Method:       "storage.Storage.ReadQC",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*ReadResponse), err
+	return gorums.QuorumCall[*ReadResponse](ctx, c.RawConfiguration, cd)
 }
 
 // WriteQC executes the Write Quorum Call on a configuration
 // of Nodes and returns true if a majority of Nodes were updated.
-func (c *Configuration) WriteQC(ctx context.Context, in *WriteRequest) (resp *WriteResponse, err error) {
+func (c *Configuration) WriteQC(ctx context.Context, in *WriteRequest) gorums.Responses[*WriteResponse] {
 	cd := gorums.QuorumCallData{
-		Message: in,
-		Method:  "storage.Storage.WriteQC",
-	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, bool) {
-		r := make(map[uint32]*WriteResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*WriteResponse)
-		}
-		return c.qspec.WriteQCQF(req.(*WriteRequest), r)
+		Message:      in,
+		Method:       "storage.Storage.WriteQC",
+		ServerStream: false,
 	}
 
-	res, err := c.RawConfiguration.QuorumCall(ctx, cd)
-	if err != nil {
-		return nil, err
-	}
-	return res.(*WriteResponse), err
+	return gorums.QuorumCall[*WriteResponse](ctx, c.RawConfiguration, cd)
 }
 
 // ReadRPC executes the Read RPC on a single Node
@@ -292,16 +236,4 @@ func RegisterStorageServer(srv *gorums.Server, impl StorageServer) {
 		defer ctx.Release()
 		impl.WriteMulticast(ctx, req)
 	})
-}
-
-type internalReadResponse struct {
-	nid   uint32
-	reply *ReadResponse
-	err   error
-}
-
-type internalWriteResponse struct {
-	nid   uint32
-	reply *WriteResponse
-	err   error
 }
