@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.5
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: correctable/correctable.proto
 
 package correctable
@@ -11,7 +11,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -23,7 +22,7 @@ const (
 )
 
 type CorrectableRequest struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -53,16 +52,23 @@ func (x *CorrectableRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use CorrectableRequest.ProtoReflect.Descriptor instead.
-func (*CorrectableRequest) Descriptor() ([]byte, []int) {
-	return file_correctable_correctable_proto_rawDescGZIP(), []int{0}
+type CorrectableRequest_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 CorrectableRequest_builder) Build() *CorrectableRequest {
+	m0 := &CorrectableRequest{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 type CorrectableResponse struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Level         int32                  `protobuf:"varint,1,opt,name=Level" json:"Level,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state            protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Level int32                  `protobuf:"varint,1,opt,name=Level"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *CorrectableResponse) Reset() {
@@ -90,59 +96,42 @@ func (x *CorrectableResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use CorrectableResponse.ProtoReflect.Descriptor instead.
-func (*CorrectableResponse) Descriptor() ([]byte, []int) {
-	return file_correctable_correctable_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *CorrectableResponse) GetLevel() int32 {
 	if x != nil {
-		return x.Level
+		return x.xxx_hidden_Level
 	}
 	return 0
 }
 
+func (x *CorrectableResponse) SetLevel(v int32) {
+	x.xxx_hidden_Level = v
+}
+
+type CorrectableResponse_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Level int32
+}
+
+func (b0 CorrectableResponse_builder) Build() *CorrectableResponse {
+	m0 := &CorrectableResponse{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Level = b.Level
+	return m0
+}
+
 var File_correctable_correctable_proto protoreflect.FileDescriptor
 
-var file_correctable_correctable_proto_rawDesc = string([]byte{
-	0x0a, 0x1d, 0x63, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x2f, 0x63, 0x6f,
-	0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
-	0x0b, 0x63, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x1a, 0x0c, 0x67, 0x6f,
-	0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x14, 0x0a, 0x12, 0x43, 0x6f,
-	0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x22, 0x2b, 0x0a, 0x13, 0x43, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x4c, 0x65, 0x76, 0x65, 0x6c,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x05, 0x4c, 0x65, 0x76, 0x65, 0x6c, 0x32, 0xc9, 0x01,
-	0x0a, 0x0f, 0x43, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x54, 0x65, 0x73,
-	0x74, 0x12, 0x56, 0x0a, 0x0b, 0x43, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65,
-	0x12, 0x1f, 0x2e, 0x63, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x2e, 0x43,
-	0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x1a, 0x20, 0x2e, 0x63, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x2e,
-	0x43, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x22, 0x04, 0xa8, 0xb5, 0x18, 0x01, 0x12, 0x5e, 0x0a, 0x11, 0x43, 0x6f, 0x72,
-	0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x12, 0x1f,
-	0x2e, 0x63, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x2e, 0x43, 0x6f, 0x72,
-	0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x20, 0x2e, 0x63, 0x6f, 0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x2e, 0x43, 0x6f,
-	0x72, 0x72, 0x65, 0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
-	0x65, 0x22, 0x04, 0xa8, 0xb5, 0x18, 0x01, 0x30, 0x01, 0x42, 0x30, 0x5a, 0x29, 0x67, 0x69, 0x74,
-	0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f,
-	0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x63, 0x6f, 0x72, 0x72, 0x65,
-	0x63, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x92, 0x03, 0x02, 0x08, 0x02, 0x62, 0x08, 0x65, 0x64, 0x69,
-	0x74, 0x69, 0x6f, 0x6e, 0x73, 0x70, 0xe8, 0x07,
-})
-
-var (
-	file_correctable_correctable_proto_rawDescOnce sync.Once
-	file_correctable_correctable_proto_rawDescData []byte
-)
-
-func file_correctable_correctable_proto_rawDescGZIP() []byte {
-	file_correctable_correctable_proto_rawDescOnce.Do(func() {
-		file_correctable_correctable_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_correctable_correctable_proto_rawDesc), len(file_correctable_correctable_proto_rawDesc)))
-	})
-	return file_correctable_correctable_proto_rawDescData
-}
+const file_correctable_correctable_proto_rawDesc = "" +
+	"\n" +
+	"\x1dcorrectable/correctable.proto\x12\vcorrectable\x1a\fgorums.proto\"\x14\n" +
+	"\x12CorrectableRequest\"+\n" +
+	"\x13CorrectableResponse\x12\x14\n" +
+	"\x05Level\x18\x01 \x01(\x05R\x05Level2\xc9\x01\n" +
+	"\x0fCorrectableTest\x12V\n" +
+	"\vCorrectable\x12\x1f.correctable.CorrectableRequest\x1a .correctable.CorrectableResponse\"\x04\xa0\xb5\x18\x01\x12^\n" +
+	"\x11CorrectableStream\x12\x1f.correctable.CorrectableRequest\x1a .correctable.CorrectableResponse\"\x04\xa0\xb5\x18\x010\x01B0Z)github.com/relab/gorums/tests/correctable\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_correctable_correctable_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_correctable_correctable_proto_goTypes = []any{
