@@ -1,93 +1,77 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.8.0-devel
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+125b0cfb
+// 	protoc            v6.30.2
 // source: correctable/correctable.proto
 
 package correctable
 
 import (
 	context "context"
-	fmt "fmt"
 	gorums "github.com/relab/gorums"
 	ordering "github.com/relab/gorums/ordering"
 	encoding "google.golang.org/grpc/encoding"
 	proto "google.golang.org/protobuf/proto"
-	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 )
 
 const (
 	// Verify that this generated code is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(8 - gorums.MinVersion)
+	_ = gorums.EnforceVersion(9 - gorums.MinVersion)
 	// Verify that the gorums runtime is sufficiently up-to-date.
-	_ = gorums.EnforceVersion(gorums.MaxVersion - 8)
+	_ = gorums.EnforceVersion(gorums.MaxVersion - 9)
 )
 
 // A Configuration represents a static set of nodes on which quorum remote
 // procedure calls may be invoked.
 type Configuration struct {
 	gorums.RawConfiguration
-	qspec QuorumSpec
-	nodes []*Node
 }
 
-func NewConfiguration(qspec QuorumSpec, cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *Configuration, err error) {
-	c = &Configuration{
-		qspec: qspec,
-	}
+// NewConfiguration returns a configuration based on the provided list of nodes (required)
+// and an optional quorum specification. The QuorumSpec is necessary for call types that
+// must process replies. For configurations only used for unicast or multicast call types,
+// a QuorumSpec is not needed.
+// Nodes can be supplied using WithNodeMap or WithNodeList.
+// Using any other type of NodeListOption will not work.
+// The ManagerOption list controls how the nodes in the configuration are created.
+func NewConfiguration(cfg gorums.NodeListOption, opts ...gorums.ManagerOption) (c *Configuration, err error) {
+	c = &Configuration{}
 	c.RawConfiguration, err = gorums.NewRawConfiguration(cfg, opts...)
 	if err != nil {
 		return nil, err
 	}
-	c.nodes = make([]*Node, c.Size())
-	for i, n := range c.RawConfiguration.RawNodes {
-		c.nodes[i] = &Node{n}
-	}
 	return c, nil
 }
 
-// ConfigurationFromRaw returns a new Configuration from the given raw configuration and QuorumSpec.
+// ConfigurationFromRaw returns a new configuration from the given raw configuration.
 //
-// This function may for example be used to "clone" a configuration but install a different QuorumSpec:
-//
-//	cfg1, err := mgr.NewConfiguration(qspec1, opts...)
-//	cfg2 := ConfigurationFromRaw(cfg1.RawConfiguration, qspec2)
-func ConfigurationFromRaw(rawCfg gorums.RawConfiguration, qspec QuorumSpec) (*Configuration, error) {
-	// return an error if the QuorumSpec interface is not empty and no implementation was provided.
-	var test interface{} = struct{}{}
-	if _, empty := test.(QuorumSpec); !empty && qspec == nil {
-		return nil, fmt.Errorf("config: missing required QuorumSpec")
-	}
+//	cfg1, err := pb.NewConfiguration(nodeList, opts...)
+//	cfg2 := ConfigurationFromRaw(cfg1.RawConfig)
+func ConfigurationFromRaw(rawCfg gorums.RawConfiguration) (*Configuration, error) {
 	newCfg := &Configuration{
 		RawConfiguration: rawCfg,
-		qspec:            qspec,
-	}
-	// initialize the nodes slice
-	newCfg.nodes = make([]*Node, newCfg.Size())
-	for i, n := range rawCfg.Nodes() {
-		newCfg.nodes[i] = &Node{n}
 	}
 	return newCfg, nil
 }
 
-func (c *Configuration) SubConfiguration(qspec QuorumSpec, cfg gorums.NodeListOption) (subCfg *Configuration, err error) {
-	subCfg = &Configuration{
-		qspec: qspec,
-	}
+// SubConfiguration allows for making a new Configuration from the ManagerOption list and
+// node list of another configuration,
+// Nodes can be supplied using WithNodeMap or WithNodeList, or WithNodeIDs.
+// A new configuration can also be created from an existing configuration,
+// using the And, WithNewNodes, Except, and WithoutNodes methods.
+func (c *Configuration) SubConfiguration(cfg gorums.NodeListOption) (subCfg *Configuration, err error) {
+	subCfg = &Configuration{}
 	subCfg.RawConfiguration, err = c.SubRawConfiguration(cfg)
 	if err != nil {
 		return nil, err
-	}
-	subCfg.nodes = make([]*Node, subCfg.Size())
-	for i, n := range subCfg.RawConfiguration.Nodes() {
-		subCfg.nodes[i] = &Node{n}
 	}
 	return subCfg, nil
 }
 
 // Close closes a configuration created from the NewConfiguration method
 //
-// NOTE: A configuration created with ConfigurationFromRaw is closed when the original configuration is closed
+// NOTE: A configuration created with ConfigurationFromRaw or SubConfiguration closes and
+// is closed when the original configuration or any of the subconfigurations are closed.
 // If you want the configurations to be independent you need to use NewConfiguration
 func (c *Configuration) Close() error {
 	return c.RawConfiguration.Close()
@@ -98,17 +82,21 @@ func (c *Configuration) Close() error {
 //
 // NOTE: mutating the returned slice is not supported.
 func (c *Configuration) Nodes() []*Node {
-	return c.nodes
+	nodes := make([]*Node, c.Size())
+	for i, n := range c.RawConfiguration.RawNodes {
+		nodes[i] = &Node{n}
+	}
+	return nodes
 }
 
 // And returns a NodeListOption that can be used to create a new configuration combining c and d.
-func (c Configuration) And(d *Configuration) gorums.NodeListOption {
+func (c *Configuration) And(d *Configuration) gorums.NodeListOption {
 	return c.RawConfiguration.And(d.RawConfiguration)
 }
 
 // Except returns a NodeListOption that can be used to create a new configuration
 // from c without the nodes in rm.
-func (c Configuration) Except(rm *Configuration) gorums.NodeListOption {
+func (c *Configuration) Except(rm *Configuration) gorums.NodeListOption {
 	return c.RawConfiguration.Except(rm.RawConfiguration)
 }
 
@@ -126,73 +114,39 @@ type Node struct {
 
 // CorrectableTestClient is the client interface for the CorrectableTest service.
 type CorrectableTestClient interface {
-	Correctable(ctx context.Context, in *CorrectableRequest) *CorrectableCorrectableResponse
-	CorrectableStream(ctx context.Context, in *CorrectableRequest) *CorrectableStreamCorrectableResponse
+	Correctable(ctx context.Context, in *CorrectableRequest) gorums.Responses[*CorrectableResponse]
+	CorrectableStream(ctx context.Context, in *CorrectableRequest) gorums.Responses[*CorrectableResponse]
 }
 
 // enforce interface compliance
 var _ CorrectableTestClient = (*Configuration)(nil)
 
-// Correctable asynchronously invokes a correctable quorum call on each node
-// in configuration c and returns a CorrectableCorrectableResponse, which can be used
-// to inspect any replies or errors when available.
-func (c *Configuration) Correctable(ctx context.Context, in *CorrectableRequest) *CorrectableCorrectableResponse {
-	cd := gorums.CorrectableCallData{
+// Correctable is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+func (c *Configuration) Correctable(ctx context.Context, in *CorrectableRequest) gorums.Responses[*CorrectableResponse] {
+	cd := gorums.QuorumCallData{
 		Message:      in,
 		Method:       "correctable.CorrectableTest.Correctable",
 		ServerStream: false,
 	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, int, bool) {
-		r := make(map[uint32]*CorrectableResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*CorrectableResponse)
-		}
-		return c.qspec.CorrectableQF(req.(*CorrectableRequest), r)
-	}
 
-	corr := c.RawConfiguration.CorrectableCall(ctx, cd)
-	return &CorrectableCorrectableResponse{corr}
+	return gorums.QuorumCall[*CorrectableResponse](ctx, c.RawConfiguration, cd)
 }
 
-// CorrectableStream asynchronously invokes a correctable quorum call on each node
-// in configuration c and returns a CorrectableStreamCorrectableResponse, which can be used
-// to inspect any replies or errors when available.
-// This method supports server-side preliminary replies (correctable stream).
-func (c *Configuration) CorrectableStream(ctx context.Context, in *CorrectableRequest) *CorrectableStreamCorrectableResponse {
-	cd := gorums.CorrectableCallData{
+// CorrectableStream is a quorum call invoked on each node in configuration c,
+
+// with the same argument in, and returns the responses as an iterator.
+
+// This is a streaming quorum call, so each can respond with any amount of responses.
+func (c *Configuration) CorrectableStream(ctx context.Context, in *CorrectableRequest) gorums.Responses[*CorrectableResponse] {
+	cd := gorums.QuorumCallData{
 		Message:      in,
 		Method:       "correctable.CorrectableTest.CorrectableStream",
 		ServerStream: true,
 	}
-	cd.QuorumFunction = func(req protoreflect.ProtoMessage, replies map[uint32]protoreflect.ProtoMessage) (protoreflect.ProtoMessage, int, bool) {
-		r := make(map[uint32]*CorrectableResponse, len(replies))
-		for k, v := range replies {
-			r[k] = v.(*CorrectableResponse)
-		}
-		return c.qspec.CorrectableStreamQF(req.(*CorrectableRequest), r)
-	}
 
-	corr := c.RawConfiguration.CorrectableCall(ctx, cd)
-	return &CorrectableStreamCorrectableResponse{corr}
-}
-
-// QuorumSpec is the interface of quorum functions for CorrectableTest.
-type QuorumSpec interface {
-	gorums.ConfigOption
-
-	// CorrectableQF is the quorum function for the Correctable
-	// correctable quorum call method. The in parameter is the request object
-	// supplied to the Correctable method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *CorrectableRequest'.
-	CorrectableQF(in *CorrectableRequest, replies map[uint32]*CorrectableResponse) (*CorrectableResponse, int, bool)
-
-	// CorrectableStreamQF is the quorum function for the CorrectableStream
-	// correctable stream quorum call method. The in parameter is the request object
-	// supplied to the CorrectableStream method at call time, and may or may not
-	// be used by the quorum function. If the in parameter is not needed
-	// you should implement your quorum function with '_ *CorrectableRequest'.
-	CorrectableStreamQF(in *CorrectableRequest, replies map[uint32]*CorrectableResponse) (*CorrectableResponse, int, bool)
+	return gorums.QuorumCall[*CorrectableResponse](ctx, c.RawConfiguration, cd)
 }
 
 // CorrectableTest is the server-side API for the CorrectableTest Service
@@ -220,46 +174,4 @@ func RegisterCorrectableTestServer(srv *gorums.Server, impl CorrectableTestServe
 			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
 		}
 	})
-}
-
-type internalCorrectableResponse struct {
-	nid   uint32
-	reply *CorrectableResponse
-	err   error
-}
-
-// CorrectableCorrectableResponse is a correctable object for processing replies.
-type CorrectableCorrectableResponse struct {
-	*gorums.Correctable
-}
-
-// Get returns the reply, level and any error associated with the
-// called method. The method does not block until a (possibly
-// intermediate) reply or error is available. Level is set to LevelNotSet if no
-// reply has yet been received. The Done or Watch methods should be used to
-// ensure that a reply is available.
-func (c *CorrectableCorrectableResponse) Get() (*CorrectableResponse, int, error) {
-	resp, level, err := c.Correctable.Get()
-	if err != nil {
-		return nil, level, err
-	}
-	return resp.(*CorrectableResponse), level, err
-}
-
-// CorrectableStreamCorrectableResponse is a correctable object for processing replies.
-type CorrectableStreamCorrectableResponse struct {
-	*gorums.Correctable
-}
-
-// Get returns the reply, level and any error associated with the
-// called method. The method does not block until a (possibly
-// intermediate) reply or error is available. Level is set to LevelNotSet if no
-// reply has yet been received. The Done or Watch methods should be used to
-// ensure that a reply is available.
-func (c *CorrectableStreamCorrectableResponse) Get() (*CorrectableResponse, int, error) {
-	resp, level, err := c.Correctable.Get()
-	if err != nil {
-		return nil, level, err
-	}
-	return resp.(*CorrectableResponse), level, err
 }
