@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.5
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: metadata/metadata.proto
 
 package metadata
@@ -11,7 +11,6 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -23,8 +22,8 @@ const (
 )
 
 type NodeID struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	ID            uint32                 `protobuf:"varint,1,opt,name=ID" json:"ID,omitempty"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_ID uint32                 `protobuf:"varint,1,opt,name=ID"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -54,23 +53,36 @@ func (x *NodeID) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use NodeID.ProtoReflect.Descriptor instead.
-func (*NodeID) Descriptor() ([]byte, []int) {
-	return file_metadata_metadata_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *NodeID) GetID() uint32 {
 	if x != nil {
-		return x.ID
+		return x.xxx_hidden_ID
 	}
 	return 0
 }
 
+func (x *NodeID) SetID(v uint32) {
+	x.xxx_hidden_ID = v
+}
+
+type NodeID_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	ID uint32
+}
+
+func (b0 NodeID_builder) Build() *NodeID {
+	m0 := &NodeID{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_ID = b.ID
+	return m0
+}
+
 type IPAddr struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Addr          string                 `protobuf:"bytes,1,opt,name=Addr" json:"Addr,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state           protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Addr string                 `protobuf:"bytes,1,opt,name=Addr"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *IPAddr) Reset() {
@@ -98,54 +110,43 @@ func (x *IPAddr) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use IPAddr.ProtoReflect.Descriptor instead.
-func (*IPAddr) Descriptor() ([]byte, []int) {
-	return file_metadata_metadata_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *IPAddr) GetAddr() string {
 	if x != nil {
-		return x.Addr
+		return x.xxx_hidden_Addr
 	}
 	return ""
 }
 
+func (x *IPAddr) SetAddr(v string) {
+	x.xxx_hidden_Addr = v
+}
+
+type IPAddr_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Addr string
+}
+
+func (b0 IPAddr_builder) Build() *IPAddr {
+	m0 := &IPAddr{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Addr = b.Addr
+	return m0
+}
+
 var File_metadata_metadata_proto protoreflect.FileDescriptor
 
-var file_metadata_metadata_proto_rawDesc = string([]byte{
-	0x0a, 0x17, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x2f, 0x6d, 0x65, 0x74, 0x61, 0x64,
-	0x61, 0x74, 0x61, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64,
-	0x61, 0x74, 0x61, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x22, 0x18, 0x0a, 0x06, 0x4e, 0x6f, 0x64, 0x65, 0x49, 0x44, 0x12, 0x0e, 0x0a, 0x02, 0x49, 0x44,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x02, 0x49, 0x44, 0x22, 0x1c, 0x0a, 0x06, 0x49, 0x50,
-	0x41, 0x64, 0x64, 0x72, 0x12, 0x12, 0x0a, 0x04, 0x41, 0x64, 0x64, 0x72, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x04, 0x41, 0x64, 0x64, 0x72, 0x32, 0x7c, 0x0a, 0x0c, 0x4d, 0x65, 0x74, 0x61,
-	0x64, 0x61, 0x74, 0x61, 0x54, 0x65, 0x73, 0x74, 0x12, 0x36, 0x0a, 0x08, 0x49, 0x44, 0x46, 0x72,
-	0x6f, 0x6d, 0x4d, 0x44, 0x12, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x10, 0x2e, 0x6d,
-	0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x2e, 0x4e, 0x6f, 0x64, 0x65, 0x49, 0x44, 0x22, 0x00,
-	0x12, 0x34, 0x0a, 0x06, 0x57, 0x68, 0x61, 0x74, 0x49, 0x50, 0x12, 0x16, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70,
-	0x74, 0x79, 0x1a, 0x10, 0x2e, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x2e, 0x49, 0x50,
-	0x41, 0x64, 0x64, 0x72, 0x22, 0x00, 0x42, 0x2d, 0x5a, 0x26, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62,
-	0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d,
-	0x73, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61,
-	0x92, 0x03, 0x02, 0x08, 0x02, 0x62, 0x08, 0x65, 0x64, 0x69, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x70,
-	0xe8, 0x07,
-})
-
-var (
-	file_metadata_metadata_proto_rawDescOnce sync.Once
-	file_metadata_metadata_proto_rawDescData []byte
-)
-
-func file_metadata_metadata_proto_rawDescGZIP() []byte {
-	file_metadata_metadata_proto_rawDescOnce.Do(func() {
-		file_metadata_metadata_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_metadata_metadata_proto_rawDesc), len(file_metadata_metadata_proto_rawDesc)))
-	})
-	return file_metadata_metadata_proto_rawDescData
-}
+const file_metadata_metadata_proto_rawDesc = "" +
+	"\n" +
+	"\x17metadata/metadata.proto\x12\bmetadata\x1a\x1bgoogle/protobuf/empty.proto\"\x18\n" +
+	"\x06NodeID\x12\x0e\n" +
+	"\x02ID\x18\x01 \x01(\rR\x02ID\"\x1c\n" +
+	"\x06IPAddr\x12\x12\n" +
+	"\x04Addr\x18\x01 \x01(\tR\x04Addr2|\n" +
+	"\fMetadataTest\x126\n" +
+	"\bIDFromMD\x12\x16.google.protobuf.Empty\x1a\x10.metadata.NodeID\"\x00\x124\n" +
+	"\x06WhatIP\x12\x16.google.protobuf.Empty\x1a\x10.metadata.IPAddr\"\x00B-Z&github.com/relab/gorums/tests/metadata\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_metadata_metadata_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_metadata_metadata_proto_goTypes = []any{
