@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.5
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: oneway/oneway.proto
 
 package oneway
@@ -11,7 +11,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -23,10 +22,10 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Num           uint64                 `protobuf:"varint,1,opt,name=Num" json:"Num,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state          protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Num uint64                 `protobuf:"varint,1,opt,name=Num"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *Request) Reset() {
@@ -54,20 +53,33 @@ func (x *Request) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Request.ProtoReflect.Descriptor instead.
-func (*Request) Descriptor() ([]byte, []int) {
-	return file_oneway_oneway_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *Request) GetNum() uint64 {
 	if x != nil {
-		return x.Num
+		return x.xxx_hidden_Num
 	}
 	return 0
 }
 
+func (x *Request) SetNum(v uint64) {
+	x.xxx_hidden_Num = v
+}
+
+type Request_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Num uint64
+}
+
+func (b0 Request_builder) Build() *Request {
+	m0 := &Request{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Num = b.Num
+	return m0
+}
+
 type Empty struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -97,47 +109,31 @@ func (x *Empty) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Empty.ProtoReflect.Descriptor instead.
-func (*Empty) Descriptor() ([]byte, []int) {
-	return file_oneway_oneway_proto_rawDescGZIP(), []int{1}
+type Empty_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 Empty_builder) Build() *Empty {
+	m0 := &Empty{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 var File_oneway_oneway_proto protoreflect.FileDescriptor
 
-var file_oneway_oneway_proto_rawDesc = string([]byte{
-	0x0a, 0x13, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2f, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x1a, 0x0c, 0x67,
-	0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x1b, 0x0a, 0x07, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x4e, 0x75, 0x6d, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x03, 0x4e, 0x75, 0x6d, 0x22, 0x07, 0x0a, 0x05, 0x45, 0x6d, 0x70, 0x74,
-	0x79, 0x32, 0xae, 0x01, 0x0a, 0x0a, 0x4f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x54, 0x65, 0x73, 0x74,
-	0x12, 0x2f, 0x0a, 0x07, 0x55, 0x6e, 0x69, 0x63, 0x61, 0x73, 0x74, 0x12, 0x0f, 0x2e, 0x6f, 0x6e,
-	0x65, 0x77, 0x61, 0x79, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0d, 0x2e, 0x6f,
-	0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x04, 0x90, 0xb5, 0x18,
-	0x01, 0x12, 0x31, 0x0a, 0x09, 0x4d, 0x75, 0x6c, 0x74, 0x69, 0x63, 0x61, 0x73, 0x74, 0x12, 0x0f,
-	0x2e, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x0d, 0x2e, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x04,
-	0x98, 0xb5, 0x18, 0x01, 0x12, 0x3c, 0x0a, 0x10, 0x4d, 0x75, 0x6c, 0x74, 0x69, 0x63, 0x61, 0x73,
-	0x74, 0x50, 0x65, 0x72, 0x4e, 0x6f, 0x64, 0x65, 0x12, 0x0f, 0x2e, 0x6f, 0x6e, 0x65, 0x77, 0x61,
-	0x79, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0d, 0x2e, 0x6f, 0x6e, 0x65, 0x77,
-	0x61, 0x79, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x08, 0x98, 0xb5, 0x18, 0x01, 0xa0, 0xb6,
-	0x18, 0x01, 0x42, 0x2b, 0x5a, 0x24, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
-	0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65,
-	0x73, 0x74, 0x73, 0x2f, 0x6f, 0x6e, 0x65, 0x77, 0x61, 0x79, 0x92, 0x03, 0x02, 0x08, 0x02, 0x62,
-	0x08, 0x65, 0x64, 0x69, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x70, 0xe8, 0x07,
-})
-
-var (
-	file_oneway_oneway_proto_rawDescOnce sync.Once
-	file_oneway_oneway_proto_rawDescData []byte
-)
-
-func file_oneway_oneway_proto_rawDescGZIP() []byte {
-	file_oneway_oneway_proto_rawDescOnce.Do(func() {
-		file_oneway_oneway_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_oneway_oneway_proto_rawDesc), len(file_oneway_oneway_proto_rawDesc)))
-	})
-	return file_oneway_oneway_proto_rawDescData
-}
+const file_oneway_oneway_proto_rawDesc = "" +
+	"\n" +
+	"\x13oneway/oneway.proto\x12\x06oneway\x1a\fgorums.proto\"\x1b\n" +
+	"\aRequest\x12\x10\n" +
+	"\x03Num\x18\x01 \x01(\x04R\x03Num\"\a\n" +
+	"\x05Empty2\xae\x01\n" +
+	"\n" +
+	"OnewayTest\x12/\n" +
+	"\aUnicast\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\x04\x90\xb5\x18\x01\x121\n" +
+	"\tMulticast\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\x04\x98\xb5\x18\x01\x12<\n" +
+	"\x10MulticastPerNode\x12\x0f.oneway.Request\x1a\r.oneway.Empty\"\b\x98\xb5\x18\x01\xa0\xb6\x18\x01B+Z$github.com/relab/gorums/tests/oneway\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_oneway_oneway_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_oneway_oneway_proto_goTypes = []any{
