@@ -2,8 +2,8 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.5
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: qf/qf.proto
 
 package qf
@@ -13,7 +13,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -25,10 +24,10 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Value         int64                  `protobuf:"varint,1,opt,name=Value" json:"Value,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state            protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Value int64                  `protobuf:"varint,1,opt,name=Value"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *Request) Reset() {
@@ -56,23 +55,36 @@ func (x *Request) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Request.ProtoReflect.Descriptor instead.
-func (*Request) Descriptor() ([]byte, []int) {
-	return file_qf_qf_proto_rawDescGZIP(), []int{0}
-}
-
 func (x *Request) GetValue() int64 {
 	if x != nil {
-		return x.Value
+		return x.xxx_hidden_Value
 	}
 	return 0
 }
 
+func (x *Request) SetValue(v int64) {
+	x.xxx_hidden_Value = v
+}
+
+type Request_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Value int64
+}
+
+func (b0 Request_builder) Build() *Request {
+	m0 := &Request{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Value = b.Value
+	return m0
+}
+
 type Response struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	Result        int64                  `protobuf:"varint,1,opt,name=Result" json:"Result,omitempty"`
-	unknownFields protoimpl.UnknownFields
-	sizeCache     protoimpl.SizeCache
+	state             protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_Result int64                  `protobuf:"varint,1,opt,name=Result"`
+	unknownFields     protoimpl.UnknownFields
+	sizeCache         protoimpl.SizeCache
 }
 
 func (x *Response) Reset() {
@@ -100,51 +112,43 @@ func (x *Response) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Response.ProtoReflect.Descriptor instead.
-func (*Response) Descriptor() ([]byte, []int) {
-	return file_qf_qf_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *Response) GetResult() int64 {
 	if x != nil {
-		return x.Result
+		return x.xxx_hidden_Result
 	}
 	return 0
 }
 
+func (x *Response) SetResult(v int64) {
+	x.xxx_hidden_Result = v
+}
+
+type Response_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	Result int64
+}
+
+func (b0 Response_builder) Build() *Response {
+	m0 := &Response{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_Result = b.Result
+	return m0
+}
+
 var File_qf_qf_proto protoreflect.FileDescriptor
 
-var file_qf_qf_proto_rawDesc = string([]byte{
-	0x0a, 0x0b, 0x71, 0x66, 0x2f, 0x71, 0x66, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x02, 0x71,
-	0x66, 0x1a, 0x0c, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22,
-	0x1f, 0x0a, 0x07, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x56, 0x61,
-	0x6c, 0x75, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x03, 0x52, 0x05, 0x56, 0x61, 0x6c, 0x75, 0x65,
-	0x22, 0x22, 0x0a, 0x08, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x16, 0x0a, 0x06,
-	0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x03, 0x52, 0x06, 0x52, 0x65,
-	0x73, 0x75, 0x6c, 0x74, 0x32, 0x69, 0x0a, 0x0e, 0x51, 0x75, 0x6f, 0x72, 0x75, 0x6d, 0x46, 0x75,
-	0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x29, 0x0a, 0x06, 0x55, 0x73, 0x65, 0x52, 0x65, 0x71,
-	0x12, 0x0b, 0x2e, 0x71, 0x66, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0c, 0x2e,
-	0x71, 0x66, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04, 0xa0, 0xb5, 0x18,
-	0x01, 0x12, 0x2c, 0x0a, 0x09, 0x49, 0x67, 0x6e, 0x6f, 0x72, 0x65, 0x52, 0x65, 0x71, 0x12, 0x0b,
-	0x2e, 0x71, 0x66, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0c, 0x2e, 0x71, 0x66,
-	0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x04, 0xa0, 0xb5, 0x18, 0x01, 0x42,
-	0x27, 0x5a, 0x20, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65,
-	0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73,
-	0x2f, 0x71, 0x66, 0x92, 0x03, 0x02, 0x08, 0x02, 0x62, 0x08, 0x65, 0x64, 0x69, 0x74, 0x69, 0x6f,
-	0x6e, 0x73, 0x70, 0xe8, 0x07,
-})
-
-var (
-	file_qf_qf_proto_rawDescOnce sync.Once
-	file_qf_qf_proto_rawDescData []byte
-)
-
-func file_qf_qf_proto_rawDescGZIP() []byte {
-	file_qf_qf_proto_rawDescOnce.Do(func() {
-		file_qf_qf_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_qf_qf_proto_rawDesc), len(file_qf_qf_proto_rawDesc)))
-	})
-	return file_qf_qf_proto_rawDescData
-}
+const file_qf_qf_proto_rawDesc = "" +
+	"\n" +
+	"\vqf/qf.proto\x12\x02qf\x1a\fgorums.proto\"\x1f\n" +
+	"\aRequest\x12\x14\n" +
+	"\x05Value\x18\x01 \x01(\x03R\x05Value\"\"\n" +
+	"\bResponse\x12\x16\n" +
+	"\x06Result\x18\x01 \x01(\x03R\x06Result2i\n" +
+	"\x0eQuorumFunction\x12)\n" +
+	"\x06UseReq\x12\v.qf.Request\x1a\f.qf.Response\"\x04\xa0\xb5\x18\x01\x12,\n" +
+	"\tIgnoreReq\x12\v.qf.Request\x1a\f.qf.Response\"\x04\xa0\xb5\x18\x01B'Z github.com/relab/gorums/tests/qf\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_qf_qf_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_qf_qf_proto_goTypes = []any{
