@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.5
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: tls/tls.proto
 
 package tls
@@ -10,7 +10,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -22,7 +21,7 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -52,14 +51,21 @@ func (x *Request) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Request.ProtoReflect.Descriptor instead.
-func (*Request) Descriptor() ([]byte, []int) {
-	return file_tls_tls_proto_rawDescGZIP(), []int{0}
+type Request_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 Request_builder) Build() *Request {
+	m0 := &Request{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 type Response struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
-	OK            bool                   `protobuf:"varint,1,opt,name=OK" json:"OK,omitempty"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
+	xxx_hidden_OK bool                   `protobuf:"varint,1,opt,name=OK"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -89,45 +95,41 @@ func (x *Response) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Response.ProtoReflect.Descriptor instead.
-func (*Response) Descriptor() ([]byte, []int) {
-	return file_tls_tls_proto_rawDescGZIP(), []int{1}
-}
-
 func (x *Response) GetOK() bool {
 	if x != nil {
-		return x.OK
+		return x.xxx_hidden_OK
 	}
 	return false
 }
 
+func (x *Response) SetOK(v bool) {
+	x.xxx_hidden_OK = v
+}
+
+type Response_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+	OK bool
+}
+
+func (b0 Response_builder) Build() *Response {
+	m0 := &Response{}
+	b, x := &b0, m0
+	_, _ = b, x
+	x.xxx_hidden_OK = b.OK
+	return m0
+}
+
 var File_tls_tls_proto protoreflect.FileDescriptor
 
-var file_tls_tls_proto_rawDesc = string([]byte{
-	0x0a, 0x0d, 0x74, 0x6c, 0x73, 0x2f, 0x74, 0x6c, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
-	0x03, 0x74, 0x6c, 0x73, 0x22, 0x09, 0x0a, 0x07, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22,
-	0x1a, 0x0a, 0x08, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x0e, 0x0a, 0x02, 0x4f,
-	0x4b, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x02, 0x4f, 0x4b, 0x32, 0x2f, 0x0a, 0x03, 0x54,
-	0x4c, 0x53, 0x12, 0x28, 0x0a, 0x07, 0x54, 0x65, 0x73, 0x74, 0x54, 0x4c, 0x53, 0x12, 0x0c, 0x2e,
-	0x74, 0x6c, 0x73, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0d, 0x2e, 0x74, 0x6c,
-	0x73, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x42, 0x28, 0x5a, 0x21,
-	0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62,
-	0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x74, 0x6c,
-	0x73, 0x92, 0x03, 0x02, 0x08, 0x02, 0x62, 0x08, 0x65, 0x64, 0x69, 0x74, 0x69, 0x6f, 0x6e, 0x73,
-	0x70, 0xe8, 0x07,
-})
-
-var (
-	file_tls_tls_proto_rawDescOnce sync.Once
-	file_tls_tls_proto_rawDescData []byte
-)
-
-func file_tls_tls_proto_rawDescGZIP() []byte {
-	file_tls_tls_proto_rawDescOnce.Do(func() {
-		file_tls_tls_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_tls_tls_proto_rawDesc), len(file_tls_tls_proto_rawDesc)))
-	})
-	return file_tls_tls_proto_rawDescData
-}
+const file_tls_tls_proto_rawDesc = "" +
+	"\n" +
+	"\rtls/tls.proto\x12\x03tls\"\t\n" +
+	"\aRequest\"\x1a\n" +
+	"\bResponse\x12\x0e\n" +
+	"\x02OK\x18\x01 \x01(\bR\x02OK2/\n" +
+	"\x03TLS\x12(\n" +
+	"\aTestTLS\x12\f.tls.Request\x1a\r.tls.Response\"\x00B(Z!github.com/relab/gorums/tests/tls\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_tls_tls_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
 var file_tls_tls_proto_goTypes = []any{
