@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.5
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: unresponsive/unresponsive.proto
 
 package unresponsive
@@ -10,7 +10,6 @@ import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
-	sync "sync"
 	unsafe "unsafe"
 )
 
@@ -22,7 +21,7 @@ const (
 )
 
 type Empty struct {
-	state         protoimpl.MessageState `protogen:"open.v1"`
+	state         protoimpl.MessageState `protogen:"opaque.v1"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -52,40 +51,26 @@ func (x *Empty) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Empty.ProtoReflect.Descriptor instead.
-func (*Empty) Descriptor() ([]byte, []int) {
-	return file_unresponsive_unresponsive_proto_rawDescGZIP(), []int{0}
+type Empty_builder struct {
+	_ [0]func() // Prevents comparability and use of unkeyed literals for the builder.
+
+}
+
+func (b0 Empty_builder) Build() *Empty {
+	m0 := &Empty{}
+	b, x := &b0, m0
+	_, _ = b, x
+	return m0
 }
 
 var File_unresponsive_unresponsive_proto protoreflect.FileDescriptor
 
-var file_unresponsive_unresponsive_proto_rawDesc = string([]byte{
-	0x0a, 0x1f, 0x75, 0x6e, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65, 0x2f, 0x75,
-	0x6e, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x12, 0x0c, 0x75, 0x6e, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65, 0x22,
-	0x07, 0x0a, 0x05, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x32, 0x4e, 0x0a, 0x0c, 0x55, 0x6e, 0x72, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65, 0x12, 0x3e, 0x0a, 0x10, 0x54, 0x65, 0x73, 0x74,
-	0x55, 0x6e, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65, 0x12, 0x13, 0x2e, 0x75,
-	0x6e, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65, 0x2e, 0x45, 0x6d, 0x70, 0x74,
-	0x79, 0x1a, 0x13, 0x2e, 0x75, 0x6e, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65,
-	0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x00, 0x42, 0x31, 0x5a, 0x2a, 0x67, 0x69, 0x74, 0x68,
-	0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61, 0x62, 0x2f, 0x67, 0x6f, 0x72,
-	0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x75, 0x6e, 0x72, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x69, 0x76, 0x65, 0x92, 0x03, 0x02, 0x08, 0x02, 0x62, 0x08, 0x65, 0x64, 0x69,
-	0x74, 0x69, 0x6f, 0x6e, 0x73, 0x70, 0xe8, 0x07,
-})
-
-var (
-	file_unresponsive_unresponsive_proto_rawDescOnce sync.Once
-	file_unresponsive_unresponsive_proto_rawDescData []byte
-)
-
-func file_unresponsive_unresponsive_proto_rawDescGZIP() []byte {
-	file_unresponsive_unresponsive_proto_rawDescOnce.Do(func() {
-		file_unresponsive_unresponsive_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_unresponsive_unresponsive_proto_rawDesc), len(file_unresponsive_unresponsive_proto_rawDesc)))
-	})
-	return file_unresponsive_unresponsive_proto_rawDescData
-}
+const file_unresponsive_unresponsive_proto_rawDesc = "" +
+	"\n" +
+	"\x1funresponsive/unresponsive.proto\x12\funresponsive\"\a\n" +
+	"\x05Empty2N\n" +
+	"\fUnresponsive\x12>\n" +
+	"\x10TestUnresponsive\x12\x13.unresponsive.Empty\x1a\x13.unresponsive.Empty\"\x00B1Z*github.com/relab/gorums/tests/unresponsive\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var file_unresponsive_unresponsive_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
 var file_unresponsive_unresponsive_proto_goTypes = []any{
